read_fonts::collections::int_set

Module bitpage

Stores a page of bits, used inside of bitset’s.

Structs

BitPage 🔒

A fixed size (512 bits wide) page of bits that records integer set membership from [0, 511].

Iter 🔒

RangeIter 🔒

Constants

ELEM_BITS 🔒

ELEM_MASK 🔒

ELEM_SIZE 🔒

PAGE_BITS 🔒

PAGE_MASK 🔒

PAGE_SIZE 🔒

Functions

elem_index_bit_mask 🔒

returns the bit to set in an element for this value

Type Aliases

Element 🔒