read_fonts::collections::int_set
Module sparse_bit_set

Provides serialization of IntSet’s to a highly compact bitset format as defined in the IFT specification:

https://w3c.github.io/IFT/Overview.html#sparse-bit-set-decoding

Structs

DecodingError

Functions

to_sparse_bit_set_with_bf

Encode this set as a sparse bit set byte encoding with a specified branch factor.