read_fonts::collections::int_set::sparse_bit_set

Function to_sparse_bit_set_with_bf

pub fn to_sparse_bit_set_with_bf<const BF: u8>(set: &IntSet<u32>) -> Vec<u8> 

Encode this set as a sparse bit set byte encoding with a specified branch factor.

Branch factor can be 2, 4, 8 or 32. It’s a compile time constant so that optimized decoding implementations can be generated by the compiler.

Sparse bit sets are a specialized, compact encoding of bit sets defined in the IFT specification: https://w3c.github.io/IFT/Overview.html#sparse-bit-set-decoding