Struct read_fonts::collections::int_set::IntSet

pub struct IntSet<T>(/* private fields */);

A fast & efficient invertible ordered set for small (up to 32-bit) unsigned integer types.

Implementations

impl IntSet<u32>

pub fn from_sparse_bit_set(data: &[u8]) -> Result<IntSet<u32>, DecodingError>

Populate this set with the values obtained from decoding the provided sparse bit set bytes.

Sparse bit sets are a specialized, compact encoding of bit sets defined in the IFT specification: https://w3c.github.io/IFT/Overview.html#sparse-bit-set-decoding

pub fn from_sparse_bit_set_bounded( data: &[u8], set_size_limit: u64, ) -> Result<IntSet<u32>, DecodingError>

Populate this set with the values obtained from decoding the provided sparse bit set bytes. The size (in number of members) of the output set is bounded to be equal to or below the provided max size. If the decoded set would contain more members than the limit an error will be returned.

Sparse bit sets are a specialized, compact encoding of bit sets defined in the IFT specification: https://w3c.github.io/IFT/Overview.html#sparse-bit-set-decoding

pub fn to_sparse_bit_set(&self) -> Vec<u8>

Encode this set as a sparse bit set byte encoding.

Sparse bit sets are a specialized, compact encoding of bit sets defined in the IFT specification: https://w3c.github.io/IFT/Overview.html#sparse-bit-set-decoding

impl<T: Domain<T>> IntSet<T>

pub fn iter(&self) -> impl DoubleEndedIterator<Item = T> + '_

Returns an iterator over all members of the set in sorted ascending order.

Note: iteration of inverted sets can be extremely slow due to the very large number of members in the set care should be taken when using .iter() in combination with an inverted set.

pub fn inclusive_iter(&self) -> Option<impl DoubleEndedIterator<Item = T> + '_>

If this is an inclusive membership set then returns an iterator over the members, otherwise returns None.

pub fn iter_after(&self, value: T) -> impl Iterator<Item = T> + '_

Returns an iterator over the members of this set that come after โvalueโ in ascending order.

Note: iteration of inverted sets can be extremely slow due to the very large number of members in the set care should be taken when using .iter() in combination with an inverted set.

pub fn iter_ranges(&self) -> impl Iterator<Item = RangeInclusive<T>> + '_

Returns an iterator over all disjoint ranges of values within the set in sorted ascending order.

pub fn insert(&mut self, val: T) -> bool

Adds a value to the set.

Returns true if the value was newly inserted.

pub fn insert_range(&mut self, range: RangeInclusive<T>)

Add all values in range as members of this set.

pub fn extend_unsorted<U: IntoIterator<Item = T>>(&mut self, iter: U)

An alternate version of extend() which is optimized for inserting an unsorted iterator of values.

pub fn remove(&mut self, val: T) -> bool

Removes a value from the set. Returns whether the value was present in the set.

pub fn remove_all<U: IntoIterator<Item = T>>(&mut self, iter: U)

pub fn remove_range(&mut self, range: RangeInclusive<T>)

Removes all values in range as members of this set.

pub fn union(&mut self, other: &IntSet<T>)

Sets the members of this set to the union of self and other.

pub fn intersect(&mut self, other: &IntSet<T>)

Sets the members of this set to the intersection of self and other.

pub fn intersects_range(&mut self, range: RangeInclusive<T>) -> bool

Returns true if this set contains at least one element in โrangeโ.

pub fn first(&self) -> Option<T>

Returns first element in the set, if any. This element is always the minimum of all elements in the set.

pub fn last(&self) -> Option<T>

Returns the last element in the set, if any. This element is always the maximum of all elements in the set.

pub fn contains(&self, val: T) -> bool

Returns true if the set contains a value.

pub fn len(&self) -> usize

Returns the number of members in this set.

pub fn is_empty(&self) -> bool

Return true if there are no members in this set.

impl<T> IntSet<T>

pub fn empty() -> IntSet<T>

Create a new empty set (inclusive).

pub fn all() -> IntSet<T>

Create a new set which contains all integers (exclusive).

pub fn is_inverted(&self) -> bool

Returns true if this set is inverted (has exclusive membership).

pub fn invert(&mut self)

Return the inverted version of this set.

pub fn clear(&mut self)

Clears the set, removing all values.