read_fonts::collections

Module range_set

Stores a disjoint collection of ranges over numeric types.

Overlapping and adjacent ranges are automatically merged together.

Structs

IntersectionIter 🔒

RangeSet

A set of disjoint ranges over numeric types.

Traits

OrdAdjacency

Allows a two values to be tested for adjacency.

Functions

range_intersection 🔒

If a and b intersect return a range representing the intersection.

range_is_subset 🔒

Returns true if the range [a_start, a_end] is a subset of [b_start, b_end].

ranges_overlap_or_adjacent 🔒

Returns true if the ranges [a_start, a_end] and [b_start, b_end] overlap or are adjacent to each other.