read_fonts::collections::range_set

Trait OrdAdjacency

pub trait OrdAdjacency {
    // Required method
    fn are_adjacent(self, rhs: Self) -> bool;
}

Allows a two values to be tested for adjacency.

Required Methods

fn are_adjacent(self, rhs: Self) -> bool

Returns true if self is adjacent on either side of rhs.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.