read_fonts
Macro be_buffer

macro_rules! be_buffer {
    ( $( $x:tt ),+ ) => { ... };
}

be_buffer!(val1, …, valn) - Constructs a BeBuffer from the provided list of values:

Values can be one of three types:

Single numeric value, ie: 10u16

Array of numeric values, ie: [1u8, 2, 3]

Tagged value: {10u16, “tag_name”}. This will insert the numeric value and attach the provide string tag to the insertion location in the buffer.