read_fonts::model

Module pen

Types for collecting the output when drawing a glyph outline.

Structs

ControlBoundsPen

Pen that generates the control bounds of a glyph outline.

NullPen

Pen that drops all drawing output into the ether.

SvgPen

Pen that generates SVG style path data.

Enums

PathElement

Single element of a path.

Traits

OutlinePen

Interface for accepting a sequence of path commands.