read_fonts::ps::cff

Struct CffFontRef

pub struct CffFontRef<'a> { /* private fields */ }

A CFF or CFF2 font.

The source data may be a raw CFF blob as embedded in a PDF or the content of a CFF or CFF2 table in an OpenType font.

Implementations

impl<'a> CffFontRef<'a>

pub fn new( data: &'a [u8], top_dict_index: u32, upem: Option<i32>, ) -> Result<Self, Error>

Creates a new font for the given CFF or CFF2 data.

Tries to determine the CFF version by reading the first word of the header.

For CFF blobs embedded in an OpenType font, the upem should be taken from the head table. Otherwise, 1000 will be assumed.

pub fn new_cff( data: &'a [u8], top_dict_index: u32, upem: Option<i32>, ) -> Result<Self, Error>

Creates a new font for the given CFF data.

For CFF blobs embedded in an OpenType font, the upem should be taken from the head table. Otherwise, 1000 will be assumed.

pub fn new_cff2(data: &'a [u8], upem: Option<i32>) -> Result<Self, Error>

Creates a new font for the given CFF2 data.

For CFF blobs embedded in an OpenType font, the upem should be taken from the head table. Otherwise, 1000 will be assumed.

pub fn data(&self) -> &'a [u8]

Returns the raw CFF blob.

pub fn version(&self) -> u16

Returns the CFF version (either 1 or 2).

pub fn metadata(&self) -> Option<Metadata<'a>>

Returns additional metadata such as font names and metrics.

pub fn is_cid(&self) -> bool

Returns true if this is a CID-keyed font.

pub fn global_subrs(&self) -> &Index<'a>

Returns the global subroutine index.

pub fn num_glyphs(&self) -> u32

Returns the number of glyphs in the CFF font.

pub fn charstrings(&self) -> &Index<'a>

Returns the charstring index.

pub fn strings(&self) -> Option<&Index<'a>>

Returns the string index.

pub fn string(&self, sid: Sid) -> Option<&'a [u8]>

Returns the string for the given identifier.

pub fn charset(&self) -> Option<Charset<'a>>

Returns the mapping for glyph identifiers.

For a CID font, this maps between CIDs and glyph identifiers. Otherwise maps between SIDs and glyph identifiers.

pub fn encoding(&self) -> Option<Encoding<'a>>

Returns the mapping from character codes to glyph identifiers.

pub fn matrix(&self) -> Option<&ScaledFontMatrix>

Returns the top level font matrix.

pub fn upem(&self) -> i32

Returns the units per em.

pub fn var_store(&self) -> Option<&ItemVariationStore<'a>>

Returns the item variation store.

Only present for CFF2 variable fonts.

pub fn num_subfonts(&self) -> u16

Returns the number of available subfonts.

pub fn subfont_index(&self, gid: GlyphId) -> Option<u16>

Returns the subfont index for the given glyph.

pub fn subfont(&self, index: u16, coords: &[F2Dot14]) -> Result<Subfont, Error>

Returns the subfont with the given index and normalized variation coordinates.

pub fn subfont_hinted( &self, index: u16, coords: &[F2Dot14], ) -> Result<(Subfont, HintingParams), Error>

Returns the subfont and hinting parameters for the given index and normalized variation coordinates.

pub fn transform(&self, subfont: &Subfont, ppem: Option<f32>) -> Transform

Returns the effective transform for the given subfont and optional size in pixels per em.

pub fn evaluate_charstring( &self, subfont: &Subfont, gid: GlyphId, coords: &[F2Dot14], sink: &mut impl CommandSink, ) -> Result<Option<Fixed>, Error>

Evaluates the charstring for the requested glyph and sends the results to the given sink.

Returns the advance with of the glyph in font units if the charstring provides one.

pub fn draw( &self, subfont: &Subfont, gid: GlyphId, coords: &[F2Dot14], ppem: Option<f32>, pen: &mut impl OutlinePen, ) -> Result<Option<f32>, Error>

Draws the glyph with an optional size in ppem to the given pen.

Returns the advance width of the glyph if the charstring provides one.