read_fonts::ps::cff

Struct Subfont

pub struct Subfont { /* private fields */ }

Metadata for a CFF subfont.

A subfont is the collection of data read from both the Font DICT and Private Dict structures. These determine the set of subroutines, metrics and hinting parameters for some group of glyphs.

Use CffFontRef::subfont_index to determine the subfont index for a particular glyph and then CffFontRef::subfont (or CffFontRef::subfont_hinted) to retrieve the associated subfont.

Implementations

impl Subfont

pub fn subrs_offset(&self) -> u32

Returns the offset to the local subroutine index from the start of the CFF blob.

pub fn default_width(&self) -> Option<Fixed>

Returns the default advance width.

The advance value for charstrings that do not contain a width.

pub fn nominal_width(&self) -> Fixed

Returns the nominal advance width.

The base advance value for charstrings that do contain a width. This should be added to the width in the charstring.

pub fn matrix(&self) -> Option<&ScaledFontMatrix>

Returns the font matrix.

pub fn vs_index(&self) -> u16

Returns the default variation store index.