read_fonts::ps::cs

Struct Evaluator

struct Evaluator<'a, S> {

    context: &'a dyn CharstringContext,
    is_type1: bool,
    blend_state: Option<BlendState<'a>>,
    sink: &'a mut S,
    is_open: bool,
    is_flexing: bool,
    seen_width_command: bool,
    have_read_width: bool,
    stem_count: usize,
    x: Fixed,
    y: Fixed,
    sbx: Fixed,
    wx: Fixed,
    stack: Stack,
    stack_ix: usize,
    in_seac: bool,
    ops_done: u32,

}

Transient state for evaluating a charstring and handling recursive subroutine calls.

Fields
context: &'a dyn CharstringContextis_type1: boolblend_state: Option<BlendState<'a>>sink: &'a mut Sis_open: boolis_flexing: bool

When the flex state is active, moveto commands simply accumulate vectors on the stack which will be used to emit curves when the flex is finalized

seen_width_command: bool

True if we’ve seen a command that might read width

have_read_width: bool

True if we’ve actually read a width

stem_count: usizex: Fixedy: Fixedsbx: Fixed

X side-bearing

wx: Fixed

X width

stack: Stackstack_ix: usizein_seac: boolops_done: u32

Number of operators or numbers processed so far

Implementations

impl<'a, S> Evaluator<'a, S>
where S: CommandSink,

fn new( context: &'a dyn CharstringContext, blend_state: Option<BlendState<'a>>, sink: &'a mut S, ) -> Self

fn evaluate(&mut self, charstring_data: &[u8]) -> Result<(), Error>

fn evaluate_impl( &mut self, charstring_data: &[u8], nesting_depth: u32, ) -> Result<bool, Error>

fn evaluate_operator( &mut self, operator: Operator, cursor: &mut Cursor<'_>, nesting_depth: u32, ) -> Result<bool, Error>

Evaluates a single charstring operator.

Returns Ok(true) if evaluation should continue.

fn read_width(&mut self) -> Result<(), Error>

fn handle_seac( &mut self, mode: SeacMode, nesting_depth: u32, ) -> Result<(), Error>

See endchar in Appendix C at https://adobe-type-tools.github.io/font-tech-notes/pdfs/5177.Type2.pdf#page=35

fn handle_flex(&mut self) -> Result<(), Error>

Emit two curves for the accumulated flex vectors.

fn handle_mm_blend( &mut self, subr_idx: i32, num_args: usize, ) -> Result<(), Error>

Handle point blending for multiple master fonts.

See https://gitlab.freedesktop.org/freetype/freetype/-/blob/80a507a6b8e3d2906ad2c8ba69329bd2fb2a85ef/src/psaux/psintrp.c#L1823

fn coords_remaining(&self) -> usize

fn ensure_open(&mut self)

fn emit_line(&mut self, x: Fixed, y: Fixed)

fn emit_curves<const N: usize>( &mut self, modes: [PointMode; N], ) -> Result<(), Error>

fn reset_stack(&mut self)