read_fonts::ps::cs

Struct NopFilterSink

pub struct NopFilterSink<'a, S> { /* private fields */ }

Command sink adapter that suppresses degenerate move and line commands.

FreeType avoids emitting empty contours and zero length lines to prevent artifacts when stem darkening is enabled. We don’t support stem darkening because it’s not enabled by any of our clients but we remove the degenerate elements regardless to match the output.

See https://gitlab.freedesktop.org/freetype/freetype/-/blob/80a507a6b8e3d2906ad2c8ba69329bd2fb2a85ef/src/psaux/pshints.c#L1786

Implementations

impl<'a, S> NopFilterSink<'a, S>
where S: CommandSink,

pub fn new(inner: &'a mut S) -> Self

Creates a new sink that suppresses degenerate move and line commands before forwarding the result to the given inner sink.