read_fonts::ps::cs

Trait CharstringContext

pub trait CharstringContext {
    // Required methods
    fn kind(&self) -> CharstringKind;
    fn seac_components(
        &self,
        base_code: i32,
        accent_code: i32,
    ) -> Result<[&[u8]; 2], Error>;
    fn global_subr(&self, index: i32) -> Result<&[u8], Error>;
    fn subr(&self, index: i32) -> Result<&[u8], Error>;

    // Provided method
    fn weight_vector(&self) -> &[Fixed] { ... }
}

Trait that provides context for charstring evaluation.

Required Methods

fn kind(&self) -> CharstringKind

Returns the type of the charstring.

fn seac_components( &self, base_code: i32, accent_code: i32, ) -> Result<[&[u8]; 2], Error>

Returns the base and accent charstrings for the seac (standard encoded accented character) operator.

fn global_subr(&self, index: i32) -> Result<&[u8], Error>

Returns the charstring for the global subroutine at the given index as encoded in the calling charstring.

fn subr(&self, index: i32) -> Result<&[u8], Error>

Returns the charstring for the local subroutine at the given index as encoded in the calling charstring.

Provided Methods

fn weight_vector(&self) -> &[Fixed]

Returns the current active weight vector for a multiple master font.