read_fonts::ps

Module num

Various types of number parsers.

Structs

BcdComponents

Components for computing a fixed point value for a binary coded decimal number.

Constants

BCD_INTEGER_LIMIT 🔒

BCD_NUMBER_LIMIT 🔒

BCD_OVERFLOW 🔒

BCD_POWER_TENS 🔒

BCD_UNDERFLOW 🔒

Functions

parse_fixed_dynamic 🔒

Parse a fixed point value with a dynamic scaling factor.

parse_int 🔒