read_fonts::ps

Module type1

Type1 fonts.

Structs

CharstringEntry 🔒

Charstrings 🔒

Collection of charstrings.

Encoding

Associates character codes with glyph names and ids.

Parser 🔒

RawDicts 🔒

Raw dictionary data for a Type1 font.

Subrs 🔒

Collection of subroutines.

Type1Font

A Type1 font.

Enums

RawEncoding 🔒

Encoding that maps characters to glyph identifiers.

Token 🔒

Constants

CHARSTRING_SEED 🔒

Decryption seed for charstring (and subroutine) data.

EEXEC_SEED 🔒

Decryption seed for eexec segment.

NOTDEF_GLYPH 🔒

Simulated .notdef glyph, same as FreeType:

PFB_BINARY_SEGMENT_TAG 🔒

PFB_TEXT_SEGMENT_TAG 🔒

Functions

decode_fixed 🔒

Decode a fixed point value, scaling to a specific power of ten.

decode_hex 🔒

Converts hex formatted data to associated bytes.

decode_int 🔒

Decode an integer, optionally with a base.

decode_int_prefix 🔒

Decode an integer at the given position, returning the value and the index of the position following the decoded integer.

decode_pfb_binary_segments 🔒

Returns an iterator over the sequence of PFB binary segments.

decode_pfb_tag 🔒

Returns the PFB tag and segment size.

decrypt 🔒

Returns an iterator yielding the decrypted bytes.

find_eexec_data 🔒

Helper to find the position of the data following the ‘eexec’ token.

is_special 🔒

Characters that always delimit tokens.

is_special_or_whitespace 🔒

is_whitespace 🔒

verify_header 🔒