read_fonts::ps::type1

Struct Parser

struct Parser<'a> {
    data: &'a [u8],
    pos: usize,
}

Fields
data: &'a [u8]pos: usize
Implementations

impl<'a> Parser<'a>

fn new(data: &'a [u8]) -> Self

fn next(&mut self) -> Option<Token<'a>>

fn accept_blob_start(&mut self) -> bool

Special lookahead case for RD or -| which indicates a binary blob in Type1 fonts.

fn accept(&mut self, token: Token<'_>) -> bool

fn expect(&mut self, token: Token<'_>) -> Option<()>

fn next_byte(&mut self) -> Option<u8>

fn peek_byte(&self) -> Option<u8>

fn read_bytes(&mut self, len: usize) -> Option<&'a [u8]>

fn skip_whitespace(&mut self) -> Option<()>

fn skip_line(&mut self)

fn skip_until(&mut self, f: impl Fn(u8) -> bool) -> usize

fn read_proc(&mut self, start: usize) -> Option<Token<'a>>

fn read_lit_string(&mut self, start: usize) -> Option<Token<'a>>

fn read_hex_string(&mut self, start: usize) -> Option<Token<'a>>

fn read_int(&mut self) -> Option<i64>

fn read_num_as_int(&mut self) -> Option<i32>

fn read_string(&mut self) -> Option<String>

impl Parser<'_>

fn read_font_matrix(&mut self) -> Option<ScaledFontMatrix>

Parse a font matrix.

Like FreeType, this is designed assuming a upem of 1000 and produces an identity matrix in that case. This is, the result is scaled such that 0.001 yields a value of 1.0.

See https://gitlab.freedesktop.org/freetype/freetype/-/blob/80a507a6b8e3d2906ad2c8ba69329bd2fb2a85ef/src/type1/t1load.c#L1403

fn read_subrs(&mut self, len_iv: i64) -> Option<Subrs>

Parse the set of subroutines.

The len_iv parameter defines the number of prefix padding bytes for encrypted data. If < 0, then the data is not encrypted.

See https://gitlab.freedesktop.org/freetype/freetype/-/blob/80a507a6b8e3d2906ad2c8ba69329bd2fb2a85ef/src/type1/t1load.c#L1720

fn read_charstrings(&mut self, len_iv: i64) -> Option<Charstrings>

Parse the set of charstrings.

The len_iv parameter defines the number of prefix padding bytes for encrypted data. If < 0, then the data is not encrypted.

See https://gitlab.freedesktop.org/freetype/freetype/-/blob/80a507a6b8e3d2906ad2c8ba69329bd2fb2a85ef/src/type1/t1load.c#L1919

fn read_font_bbox(&mut self) -> Option<[Fixed; 4]>

fn read_weight_vector(&mut self) -> Option<Vec<Fixed>>

fn read_encoding(&mut self, charstrings: &Charstrings) -> Option<RawEncoding>

Parse the encoding.

See https://gitlab.freedesktop.org/freetype/freetype/-/blob/80a507a6b8e3d2906ad2c8ba69329bd2fb2a85ef/src/type1/t1load.c#L1474

fn read_dense_encoding(&mut self, f: impl FnMut(i64, &str)) -> Option<()>

Returns a custom encoding defined by an array of /<name>, where the index represents the character code, and invokes the given callback for each.

fn read_sparse_encoding(&mut self, f: impl FnMut(i64, &str)) -> Option<()>

Reads a custom encoding defined by a map of <charcode> /<name> and invokes the given callback for each.