read_fonts::ps::type1

Struct Type1Font

pub struct Type1Font { /* private fields */ }

A Type1 font.

Implementations

impl Type1Font

pub fn new(data: &[u8]) -> Result<Self, Error>

Creates a new Type1 font from the given data.

pub fn name(&self) -> Option<&str>

Returns the PostScript name.

pub fn full_name(&self) -> Option<&str>

Returns the full font name.

pub fn family_name(&self) -> Option<&str>

Returns the font family name.

pub fn weight(&self) -> Option<&str>

Returns the weight or style name.

pub fn italic_angle(&self) -> i32

Returns the italic angle.

pub fn is_fixed_pitch(&self) -> bool

Returns true if the glyphs in this font have the same width.

pub fn underline_position(&self) -> i32

Returns the position of the top of an underline decoration.

pub fn underline_thickness(&self) -> i32

Returns the suggested size for an underline decoration.

pub fn bbox(&self) -> BoundingBox<Fixed>

Returns the font bounding box.

pub fn num_glyphs(&self) -> u32

Returns the number of glyphs in the Type1 font.

pub fn upem(&self) -> i32

Returns the units per em.

pub fn matrix(&self) -> FontMatrix

Returns the top level font matrix.

pub fn transform(&self, ppem: Option<f32>) -> Transform

Returns the appropriate transform for adjusting points and metrics.

pub fn encoding(&self) -> Option<Encoding<'_>>

Returns the character encoding.

pub fn unicode_charmap(&self) -> &Charmap

Returns the Unicode charmap for this font.

Note that this is an empty mapping if the agl feature is not enabled.

pub fn glyph_name(&self, gid: GlyphId) -> Option<&str>

Returns the glyph name for the given id.

pub fn glyph_names(&self) -> impl Iterator<Item = (GlyphId, &str)>

Returns an iterator over the pairs of glyph ids and associated names in the Type1 font.

pub fn remapped_gid(&self, original_gid: GlyphId) -> GlyphId

Given a glyph identifier in the original glyph order, returns the possibly remapped identifier.

This occurs if we remap or synthesize a .notdef glyph.

pub fn evaluate_charstring( &self, gid: GlyphId, sink: &mut impl CommandSink, ) -> Result<Option<Fixed>, Error>

Evaluates the charstring for the requested glyph and sends the results to the given sink.

Returns the advance with of the glyph in font units if the charstring provides one.

pub fn draw( &self, gid: GlyphId, ppem: Option<f32>, pen: &mut impl OutlinePen, ) -> Result<Option<f32>, Error>

Draws the glyph with an optional size in ppem to the given pen.

Returns the advance width of the glyph if the charstring provides one.