read_fonts

Module read

Traits for interpreting font data

Enums

ReadError

An error that occurs when reading font data

Traits

ComputeSize

A type that can compute its size at runtime, based on some input.

Discriminant

A trait for tables that contain offsets to subtables of heterogeneous types.

FontRead

A type that can be read from raw table data.

Format

A trait for tables that have multiple possible formats.

ReadArgs

A trait for a type that needs additional arguments to be read.

VarSize

A trait for types that have variable length.