read_fonts::read

Trait ReadArgs

pub trait ReadArgs {
    type Args: Copy;
}

A trait for a type that needs additional arguments to be read.

Types that do not require any external state use () as their args.

This is separate from FontRead so that it can also be a supertrait of ComputeSize, which does not need a lifetime.

Required Associated Types

type Args: Copy