read_fonts

Struct ArrayOfOffsets

pub struct ArrayOfOffsets<'a, T: ReadArgs, O: Scalar = Offset16> { /* private fields */ }

An array of offsets that can be resolved on access.

This bundles up the raw offsets with the data used to resolve them, along with any arguments needed to resolve those offsets; it provides a simple ergonomic interface that unburdens the user from needing to manually determine the appropriate input data and arguments for a raw offset.

Implementations

impl<'a, T, O> ArrayOfOffsets<'a, T, O>
where O: Scalar + Offset, T: ReadArgs + FontReadWithArgs<'a>, T::Args: Copy + 'static,

pub fn len(&self) -> usize

The number of offsets in the array

pub fn is_empty(&self) -> bool

true if the array is empty

pub fn get(&self, idx: usize) -> Result<T, ReadError>

Resolve the offset at the provided index.

Note: if the index is invalid this will return the InvalidCollectionIndex error variant instead of None.

pub fn iter(&self) -> impl Iterator<Item = Result<T, ReadError>> + 'a

Iterate over all of the offset targets.

Each offset will be resolved as it is encountered.