Struct read_fonts::FontRef

pub struct FontRef<'a> {
    pub table_directory: TableDirectory<'a>,
    /* private fields */
}

Reference to an in-memory font.

This is a simple implementation of the TableProvider trait backed by a borrowed slice containing font data.

Fields
table_directory: TableDirectory<'a>
Implementations

impl<'a> FontRef<'a>

pub fn new(data: &'a [u8]) -> Result<Self, ReadError>

Creates a new reference to an in-memory font backed by the given data.

The data must be a single font (not a font collection) and must begin with a table directory to be considered valid.

To load a font from a font collection, use FontRef::from_index instead.

pub fn from_index(data: &'a [u8], index: u32) -> Result<Self, ReadError>

Creates a new reference to an in-memory font at the specified index backed by the given data.

The data slice must begin with either a table directory or a ttc header to be considered valid.

In other words, this accepts either font collection (ttc) or single font (ttf/otf) files. If a single font file is provided, the index parameter must be 0.

pub fn table_data(&self, tag: Tag) -> Option<FontData<'a>>

Returns the data for the table with the specified tag, if present.