Struct read_fonts::tables::aat::StateTable

pub struct StateTable<'a> { /* private fields */ }

Implementations

impl<'a> StateTable<'a>

pub fn class(&self, glyph_id: GlyphId16) -> Result<u8, ReadError>

Returns the class table entry for the given glyph identifier.

pub fn entry(&self, state: u16, class: u8) -> Result<StateEntry, ReadError>

Returns the entry for the given state and class.