read_fonts::tables::ankr

Type Alias Ankr

pub type Ankr<'a> = TableRef<'a, AnkrMarker>;

The anchor point table.

Aliased Type

struct Ankr<'a> { /* private fields */ }

Implementations

impl<'a> Ankr<'a>

pub fn version(&self) -> u16

Version number (set to zero).

pub fn flags(&self) -> u16

Flags (currently unused; set to zero).

pub fn lookup_table_offset(&self) -> Offset32

Offset to the table’s lookup table; currently this is always 0x0000000C.

Lookup values are two byte offsets into the glyph data table.

pub fn lookup_table(&self) -> Result<LookupU16<'a>, ReadError>

Attempt to resolve lookup_table_offset.

pub fn glyph_data_table_offset(&self) -> u32

Offset to the glyph data table.

impl<'a> Ankr<'a>

pub fn anchor_points( &self, glyph_id: GlyphId, ) -> Result<&'a [AnchorPoint], ReadError>

Returns the set of anchor points for the given glyph.