Type Alias read_fonts::tables::base::Base

pub type Base<'a> = TableRef<'a, BaseMarker>;

The BASE (Baseline) table

Aliased Type

struct Base<'a> { /* private fields */ }

Implementations

impl<'a> Base<'a>

pub fn version(&self) -> MajorMinor

(major, minor) Version for the BASE table (1,0) or (1,1)

pub fn horiz_axis_offset(&self) -> Nullable<Offset16>

Offset to horizontal Axis table, from beginning of BASE table (may be NULL)

pub fn horiz_axis(&self) -> Option<Result<Axis<'a>, ReadError>>

Attempt to resolve horiz_axis_offset.

pub fn vert_axis_offset(&self) -> Nullable<Offset16>

Offset to vertical Axis table, from beginning of BASE table (may be NULL)

pub fn vert_axis(&self) -> Option<Result<Axis<'a>, ReadError>>

Attempt to resolve vert_axis_offset.

pub fn item_var_store_offset(&self) -> Option<Nullable<Offset32>>

Offset to Item Variation Store table, from beginning of BASE table (may be null)

pub fn item_var_store( &self, ) -> Option<Result<ItemVariationStore<'a>, ReadError>>

Attempt to resolve item_var_store_offset.