read_fonts::tables::base

Type Alias MinMax

pub type MinMax<'a> = TableRef<'a, MinMaxMarker>;

MinMax table

Aliased Type

struct MinMax<'a> { /* private fields */ }

Implementations

impl<'a> MinMax<'a>

pub fn min_coord_offset(&self) -> Nullable<Offset16>

Offset to BaseCoord table that defines the minimum extent value, from the beginning of MinMax table (may be NULL)

pub fn min_coord(&self) -> Option<Result<BaseCoord<'a>, ReadError>>

Attempt to resolve min_coord_offset.

pub fn max_coord_offset(&self) -> Nullable<Offset16>

Offset to BaseCoord table that defines maximum extent value, from the beginning of MinMax table (may be NULL)

pub fn max_coord(&self) -> Option<Result<BaseCoord<'a>, ReadError>>

Attempt to resolve max_coord_offset.

pub fn feat_min_max_count(&self) -> u16

Number of FeatMinMaxRecords — may be zero (0)

pub fn feat_min_max_records(&self) -> &'a [FeatMinMaxRecord]

Array of FeatMinMaxRecords, in alphabetical order by featureTableTag