read_fonts::tables::bitmap

Struct BitmapSize

#[repr(C, packed(1))]
pub struct BitmapSize {
    pub index_subtable_array_offset: BigEndian<u32>,
    pub index_tables_size: BigEndian<u32>,
    pub number_of_index_subtables: BigEndian<u32>,
    pub color_ref: BigEndian<u32>,
    pub hori: SbitLineMetrics,
    pub vert: SbitLineMetrics,
    pub start_glyph_index: BigEndian<GlyphId16>,
    pub end_glyph_index: BigEndian<GlyphId16>,
    pub ppem_x: u8,
    pub ppem_y: u8,
    pub bit_depth: u8,
    pub flags: BigEndian<BitmapFlags>,
}

BitmapSize record.

Fields
index_subtable_array_offset: BigEndian<u32>

Offset to index subtable from beginning of EBLC/CBLC.

index_tables_size: BigEndian<u32>

Number of bytes in corresponding index subtables and array.

number_of_index_subtables: BigEndian<u32>

There is an index subtable for each range or format change.

color_ref: BigEndian<u32>

Not used; set to 0.

hori: SbitLineMetrics

Line metrics for text rendered horizontally.

vert: SbitLineMetrics

Line metrics for text rendered vertically.

start_glyph_index: BigEndian<GlyphId16>

Lowest glyph index for this size.

end_glyph_index: BigEndian<GlyphId16>

Highest glyph index for this size.

ppem_x: u8

Horizontal pixels per em.

ppem_y: u8

Vertical pixels per em.

bit_depth: u8

The Microsoft rasterizer v.1.7 or greater supports the following bitDepth values, as described below: 1, 2, 4, and 8 (and 32 for CBLC).

flags: BigEndian<BitmapFlags>

Vertical or horizontal.

Implementations

impl BitmapSize

pub fn index_subtable_array_offset(&self) -> u32

Offset to index subtable from beginning of EBLC/CBLC.

pub fn index_tables_size(&self) -> u32

Number of bytes in corresponding index subtables and array.

pub fn number_of_index_subtables(&self) -> u32

There is an index subtable for each range or format change.

pub fn color_ref(&self) -> u32

Not used; set to 0.

pub fn hori(&self) -> &SbitLineMetrics

Line metrics for text rendered horizontally.

pub fn vert(&self) -> &SbitLineMetrics

Line metrics for text rendered vertically.

pub fn start_glyph_index(&self) -> GlyphId16

Lowest glyph index for this size.

pub fn end_glyph_index(&self) -> GlyphId16

Highest glyph index for this size.

pub fn ppem_x(&self) -> u8

Horizontal pixels per em.

pub fn ppem_y(&self) -> u8

Vertical pixels per em.

pub fn bit_depth(&self) -> u8

The Microsoft rasterizer v.1.7 or greater supports the following bitDepth values, as described below: 1, 2, 4, and 8 (and 32 for CBLC).

pub fn flags(&self) -> BitmapFlags

Vertical or horizontal.

impl BitmapSize

pub fn location( &self, offset_data: FontData<'_>, glyph_id: GlyphId, ) -> Result<BitmapLocation, ReadError>

Returns the bitmap location information for the given glyph.

The offset_data parameter is provided by the offset_data() method of the parent Eblc or Cblc table.

The resulting BitmapLocation value is used by the data() method in the associated Ebdt or Cbdt table to extract the bitmap data.