read_fonts::tables::colr

Type Alias PaintSkew

pub type PaintSkew<'a> = TableRef<'a, PaintSkewMarker>;

PaintSkew table

Aliased Type

struct PaintSkew<'a> { /* private fields */ }

Implementations

impl<'a> PaintSkew<'a>

pub fn format(&self) -> u8

Set to 28.

pub fn paint_offset(&self) -> Offset24

Offset to a Paint subtable.

pub fn paint(&self) -> Result<Paint<'a>, ReadError>

Attempt to resolve paint_offset.

pub fn x_skew_angle(&self) -> F2Dot14

Angle of skew in the direction of the x-axis, 180° in counter-clockwise degrees per 1.0 of value.

pub fn y_skew_angle(&self) -> F2Dot14

Angle of skew in the direction of the y-axis, 180° in counter-clockwise degrees per 1.0 of value.