read_fonts::tables

Module fvar

The Font Variations table

Modules

instance_record 🔒

An fvar InstanceRecord

Structs

AxisInstanceArrays

Shim table to handle combined axis and instance arrays.

Fvar

The fvar (Font Variations) table

InstanceRecord

The InstanceRecord

VariationAxisRecord

The VariationAxisRecord

Constants

MAX_INLINE_AVAR2_AXES 🔒

MAX_INLINE_NORMALIZE_AXES 🔒

MAX_NORMALIZE_QUADRATIC_AXES 🔒

Maximum number of axes for which we will use a quadratic path to normalize user coordinates.

Functions

apply_avar2_delta 🔒

apply_avar_mappings 🔒

normalize_user_coords 🔒

round_f64_to_i32 🔒

to_normalized_coords 🔒