Type Alias read_fonts::tables::fvar::AxisInstanceArrays

pub type AxisInstanceArrays<'a> = TableRef<'a, AxisInstanceArraysMarker>;

Shim table to handle combined axis and instance arrays.

Aliased Type

struct AxisInstanceArrays<'a> { /* private fields */ }

Implementations

impl<'a> AxisInstanceArrays<'a>

pub fn read( data: FontData<'a>, axis_count: u16, instance_count: u16, instance_size: u16 ) -> Result<Self, ReadError>

A constructor that requires additional arguments.

This type requires some external state in order to be parsed.

impl<'a> AxisInstanceArrays<'a>

pub fn axes(&self) -> &'a [VariationAxisRecord]

Variation axis record array.

pub fn instances(&self) -> ComputedArray<'a, InstanceRecord<'a>>

Instance record array.