Type Alias read_fonts::tables::gdef::Gdef

pub type Gdef<'a> = TableRef<'a, GdefMarker>;

GDEF 1.0

Aliased Type

struct Gdef<'a> { /* private fields */ }

Implementations

impl<'a> Gdef<'a>

pub fn version(&self) -> MajorMinor

The major/minor version of the GDEF table

pub fn glyph_class_def_offset(&self) -> Nullable<Offset16>

Offset to class definition table for glyph type, from beginning of GDEF header (may be NULL)

pub fn glyph_class_def(&self) -> Option<Result<ClassDef<'a>, ReadError>>

Attempt to resolve glyph_class_def_offset.

pub fn attach_list_offset(&self) -> Nullable<Offset16>

Offset to attachment point list table, from beginning of GDEF header (may be NULL)

pub fn attach_list(&self) -> Option<Result<AttachList<'a>, ReadError>>

Attempt to resolve attach_list_offset.

pub fn lig_caret_list_offset(&self) -> Nullable<Offset16>

Offset to ligature caret list table, from beginning of GDEF header (may be NULL)

pub fn lig_caret_list(&self) -> Option<Result<LigCaretList<'a>, ReadError>>

Attempt to resolve lig_caret_list_offset.

pub fn mark_attach_class_def_offset(&self) -> Nullable<Offset16>

Offset to class definition table for mark attachment type, from beginning of GDEF header (may be NULL)

pub fn mark_attach_class_def(&self) -> Option<Result<ClassDef<'a>, ReadError>>

Attempt to resolve mark_attach_class_def_offset.

pub fn mark_glyph_sets_def_offset(&self) -> Option<Nullable<Offset16>>

Offset to the table of mark glyph set definitions, from beginning of GDEF header (may be NULL)

pub fn mark_glyph_sets_def( &self, ) -> Option<Result<MarkGlyphSets<'a>, ReadError>>

Attempt to resolve mark_glyph_sets_def_offset.

pub fn item_var_store_offset(&self) -> Option<Nullable<Offset32>>

Offset to the Item Variation Store table, from beginning of GDEF header (may be NULL)

pub fn item_var_store( &self, ) -> Option<Result<ItemVariationStore<'a>, ReadError>>

Attempt to resolve item_var_store_offset.