Module read_fonts::tables::glyf

The glyf (Glyph Data) table

Modules

bytecode

TrueType hinting bytecode.

Structs

Component

A reference to another glyph. Part of CompositeGlyph.

CompositeGlyphFlags

Flags used in CompositeGlyph

CurvePoint

Point with an associated on-curve flag in a simple glyph.

PointFlags

Flags describing the properties of a point.

PointMarker

Marker bits for point flags that are set during variation delta processing and hinting.

SimpleGlyphFlags

Flags used in SimpleGlyph

Transform

Transform for a composite component.

Enums

Anchor

Anchor position for a composite component.

Glyph

Simple or composite glyph.

ToPathError

Errors that can occur when converting an outline to a path.

Functions

to_path

Converts a glyf outline described by points, flags and contour end points to a sequence of path elements and invokes the appropriate callback on the given sink for each.

Type Aliases

CompositeGlyph

CompositeGlyph

Glyf

The glyf (Glyph Data) table

SimpleGlyph

The Glyph Header