read_fonts::tables::glyf::bytecode

Module opcode

TrueType hinting opcodes.

Enums

Opcode

Operation code for a TrueType instruction.

Constants

OPCODE_LENGTHS 🔒

Size in bytes of an instruction.

OPCODE_NAMES 🔒

There doesn’t seem to be any prevailing set of mnemonics for these instructions. These are pulled from FreeType with the justification that diffing FreeType hinting traces with our own is the most efficient way to track down discrepancies. https://gitlab.freedesktop.org/freetype/freetype/-/blob/57617782464411201ce7bbc93b086c1b4d7d84a5/src/truetype/ttinterp.c#L837