Function read_fonts::tables::glyf::to_path

pub fn to_path<C: PointCoord>(
    points: &[Point<C>],
    flags: &[PointFlags],
    contours: &[u16],
    path_style: ToPathStyle,
    pen: &mut impl Pen,
) -> Result<(), ToPathError>

Converts a glyf outline described by points, flags and contour end points to a sequence of path elements and invokes the appropriate callback on the given pen for each.

The input points can have any coordinate type that implements PointCoord. Output points are always generated in f32.

This is roughly equivalent to FT_Outline_Decompose.