Struct read_fonts::tables::glyf::CompositeGlyphFlags

pub struct CompositeGlyphFlags { /* private fields */ }

Flags used in CompositeGlyph

Implementations

impl CompositeGlyphFlags

pub const ARG_1_AND_2_ARE_WORDS: Self = _

Bit 0: If this is set, the arguments are 16-bit (uint16 or int16); otherwise, they are bytes (uint8 or int8).

pub const ARGS_ARE_XY_VALUES: Self = _

Bit 1: If this is set, the arguments are signed xy values; otherwise, they are unsigned point numbers.

pub const ROUND_XY_TO_GRID: Self = _

Bit 2: If set and ARGS_ARE_XY_VALUES is also set, the xy values are rounded to the nearest grid line. Ignored if ARGS_ARE_XY_VALUES is not set.

pub const WE_HAVE_A_SCALE: Self = _

Bit 3: This indicates that there is a simple scale for the component. Otherwise, scale = 1.0.

pub const MORE_COMPONENTS: Self = _

Bit 5: Indicates at least one more glyph after this one.

pub const WE_HAVE_AN_X_AND_Y_SCALE: Self = _

Bit 6: The x direction will use a different scale from the y direction.

pub const WE_HAVE_A_TWO_BY_TWO: Self = _

Bit 7: There is a 2 by 2 transformation that will be used to scale the component.

pub const WE_HAVE_INSTRUCTIONS: Self = _

Bit 8: Following the last component are instructions for the composite character.

pub const USE_MY_METRICS: Self = _

Bit 9: If set, this forces the aw and lsb (and rsb) for the composite to be equal to those from this component glyph. This works for hinted and unhinted glyphs.

pub const OVERLAP_COMPOUND: Self = _

Bit 10: If set, the components of the compound glyph overlap. Use of this flag is not required in OpenType — that is, it is valid to have components overlap without having this flag set. It may affect behaviors in some platforms, however. (See Apple’s specification for details regarding behavior in Apple platforms.) When used, it must be set on the flag word for the first component. See additional remarks, above, for the similar OVERLAP_SIMPLE flag used in simple-glyph descriptions.

pub const SCALED_COMPONENT_OFFSET: Self = _

Bit 11: The composite is designed to have the component offset scaled. Ignored if ARGS_ARE_XY_VALUES is not set.

pub const UNSCALED_COMPONENT_OFFSET: Self = _

Bit 12: The composite is designed not to have the component offset scaled. Ignored if ARGS_ARE_XY_VALUES is not set.

impl CompositeGlyphFlags

pub const fn empty() -> Self

Returns an empty set of flags.

pub const fn all() -> Self

Returns the set containing all flags.

pub const fn bits(&self) -> u16

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u16) -> Option<Self>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u16) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn intersects(&self, other: Self) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Self) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Self)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Self)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Self)

Toggles the specified flags in-place.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: Self) -> Self

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both.

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: Self) -> Self

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the “bit-clear” operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.