Type Alias read_fonts::tables::glyf::CompositeGlyph

pub type CompositeGlyph<'a> = TableRef<'a, CompositeGlyphMarker>;

CompositeGlyph

Aliased Type

struct CompositeGlyph<'a> { /* private fields */ }

Implementations

impl<'a> CompositeGlyph<'a>

pub fn number_of_contours(&self) -> i16

If the number of contours is greater than or equal to zero, this is a simple glyph. If negative, this is a composite glyph — the value -1 should be used for composite glyphs.

pub fn x_min(&self) -> i16

Minimum x for coordinate data.

pub fn y_min(&self) -> i16

Minimum y for coordinate data.

pub fn x_max(&self) -> i16

Maximum x for coordinate data.

pub fn y_max(&self) -> i16

Maximum y for coordinate data.

pub fn component_data(&self) -> &'a [u8]

component flag glyph index of component

impl<'a> CompositeGlyph<'a>

pub fn components(&self) -> impl Iterator<Item = Component> + 'a + Clone

Returns an iterator over the components of the composite glyph.

pub fn component_glyphs_and_flags( &self ) -> impl Iterator<Item = (GlyphId, CompositeGlyphFlags)> + 'a + Clone

Returns an iterator that yields the glyph identifier and flags of each component in the composite glyph.

pub fn count_and_instructions(&self) -> (usize, Option<&'a [u8]>)

Returns the component count and TrueType interpreter instructions in a single pass.

pub fn instructions(&self) -> Option<&'a [u8]>

Returns the TrueType interpreter instructions.