Type Alias read_fonts::tables::glyf::SimpleGlyph

pub type SimpleGlyph<'a> = TableRef<'a, SimpleGlyphMarker>;

The Glyph Header

Aliased Type

struct SimpleGlyph<'a> { /* private fields */ }

Implementations

impl<'a> SimpleGlyph<'a>

pub fn number_of_contours(&self) -> i16

If the number of contours is greater than or equal to zero, this is a simple glyph. If negative, this is a composite glyph — the value -1 should be used for composite glyphs.

pub fn x_min(&self) -> i16

Minimum x for coordinate data.

pub fn y_min(&self) -> i16

Minimum y for coordinate data.

pub fn x_max(&self) -> i16

Maximum x for coordinate data.

pub fn y_max(&self) -> i16

Maximum y for coordinate data.

pub fn end_pts_of_contours(&self) -> &'a [BigEndian<u16>]

Array of point indices for the last point of each contour, in increasing numeric order

pub fn instruction_length(&self) -> u16

Total number of bytes for instructions. If instructionLength is zero, no instructions are present for this glyph, and this field is followed directly by the flags field.

pub fn instructions(&self) -> &'a [u8]

Array of instruction byte code for the glyph.

pub fn glyph_data(&self) -> &'a [u8]

the raw data for flags & x/y coordinates

impl<'a> SimpleGlyph<'a>

pub fn num_points(&self) -> usize

Returns the total number of points.

pub fn has_overlapping_contours(&self) -> bool

Returns true if the contours in the simple glyph may overlap.

pub fn read_points_fast( &self, points: &mut [Point<i32>], flags: &mut [PointFlags] ) -> Result<(), ReadError>

Reads points and flags into the provided buffers.

Drops all flag bits except on-curve. The lengths of the buffers must be equal to the value returned by num_points.

Performance

As the name implies, this is faster than using the iterator returned by points so should be used when it is possible to preallocate buffers.

pub fn points(&self) -> impl Iterator<Item = CurvePoint> + 'a + Clone

Returns an iterator over the points in the glyph.

Performance

This is slower than read_points_fast but provides access to the points without requiring a preallocated buffer.