Struct read_fonts::tables::gpos::PairValueRecord

pub struct PairValueRecord {
    pub second_glyph: BigEndian<GlyphId>,
    pub value_record1: ValueRecord,
    pub value_record2: ValueRecord,
}

Part of PairSet

Fields
second_glyph: BigEndian<GlyphId>

Glyph ID of second glyph in the pair (first glyph is listed in the Coverage table).

value_record1: ValueRecord

Positioning data for the first glyph in the pair.

value_record2: ValueRecord

Positioning data for the second glyph in the pair.

Implementations

impl PairValueRecord

pub fn second_glyph(&self) -> GlyphId

Glyph ID of second glyph in the pair (first glyph is listed in the Coverage table).

pub fn value_record1(&self) -> &ValueRecord

Positioning data for the first glyph in the pair.

pub fn value_record2(&self) -> &ValueRecord

Positioning data for the second glyph in the pair.

impl<'a> PairValueRecord

pub fn read( data: FontData<'a>, value_format1: ValueFormat, value_format2: ValueFormat ) -> Result<Self, ReadError>

A constructor that requires additional arguments.

This type requires some external state in order to be parsed.