Struct read_fonts::tables::gpos::ValueRecord

pub struct ValueRecord {
    pub x_placement: Option<BigEndian<i16>>,
    pub y_placement: Option<BigEndian<i16>>,
    pub x_advance: Option<BigEndian<i16>>,
    pub y_advance: Option<BigEndian<i16>>,
    pub x_placement_device: BigEndian<Nullable<Offset16>>,
    pub y_placement_device: BigEndian<Nullable<Offset16>>,
    pub x_advance_device: BigEndian<Nullable<Offset16>>,
    pub y_advance_device: BigEndian<Nullable<Offset16>>,
    /* private fields */
}

A Positioning ValueRecord.

NOTE: we create these manually, since parsing is weird and depends on the associated valueformat. That said, this isnโt a great representation? we could definitely do something much more in the zero-copy mode..

Fields
x_placement: Option<BigEndian<i16>>y_placement: Option<BigEndian<i16>>x_advance: Option<BigEndian<i16>>y_advance: Option<BigEndian<i16>>x_placement_device: BigEndian<Nullable<Offset16>>y_placement_device: BigEndian<Nullable<Offset16>>x_advance_device: BigEndian<Nullable<Offset16>>y_advance_device: BigEndian<Nullable<Offset16>>
Implementations

impl ValueRecord

pub fn read(data: FontData<'_>, format: ValueFormat) -> Result<Self, ReadError>

pub fn x_placement(&self) -> Option<i16>

pub fn y_placement(&self) -> Option<i16>

pub fn x_advance(&self) -> Option<i16>

pub fn y_advance(&self) -> Option<i16>

pub fn x_placement_device<'a>( &self, data: FontData<'a>, ) -> Option<Result<DeviceOrVariationIndex<'a>, ReadError>>

pub fn y_placement_device<'a>( &self, data: FontData<'a>, ) -> Option<Result<DeviceOrVariationIndex<'a>, ReadError>>

pub fn x_advance_device<'a>( &self, data: FontData<'a>, ) -> Option<Result<DeviceOrVariationIndex<'a>, ReadError>>

pub fn y_advance_device<'a>( &self, data: FontData<'a>, ) -> Option<Result<DeviceOrVariationIndex<'a>, ReadError>>