read_fonts::tables::gpos

Type Alias AnchorFormat3

pub type AnchorFormat3<'a> = TableRef<'a, AnchorFormat3Marker>;

Anchor Table Format 3: Design Units Plus Device or VariationIndex Tables

Aliased Type

struct AnchorFormat3<'a> { /* private fields */ }

Implementations

impl<'a> AnchorFormat3<'a>

pub fn anchor_format(&self) -> u16

Format identifier, = 3

pub fn x_coordinate(&self) -> i16

Horizontal value, in design units

pub fn y_coordinate(&self) -> i16

Vertical value, in design units

pub fn x_device_offset(&self) -> Nullable<Offset16>

Offset to Device table (non-variable font) / VariationIndex table (variable font) for X coordinate, from beginning of Anchor table (may be NULL)

pub fn x_device(&self) -> Option<Result<DeviceOrVariationIndex<'a>, ReadError>>

Attempt to resolve x_device_offset.

pub fn y_device_offset(&self) -> Nullable<Offset16>

Offset to Device table (non-variable font) / VariationIndex table (variable font) for Y coordinate, from beginning of Anchor table (may be NULL)

pub fn y_device(&self) -> Option<Result<DeviceOrVariationIndex<'a>, ReadError>>

Attempt to resolve y_device_offset.