read_fonts::tables::gsub

Type Alias AlternateSubstFormat1

pub type AlternateSubstFormat1<'a> = TableRef<'a, AlternateSubstFormat1Marker>;

Alternate Substitution Format 1

Aliased Type

pub struct AlternateSubstFormat1<'a> {
    pub(crate) shape: AlternateSubstFormat1Marker,
    pub(crate) data: FontData<'a>,
}

Fields
shape: AlternateSubstFormat1Markerdata: FontData<'a>
Implementations

impl<'a> AlternateSubstFormat1<'a>

pub fn subst_format(&self) -> u16

Format identifier: format = 1

pub fn coverage_offset(&self) -> Offset16

Offset to Coverage table, from beginning of substitution subtable

pub fn coverage(&self) -> Result<CoverageTable<'a>, ReadError>

Attempt to resolve coverage_offset.

pub fn alternate_set_count(&self) -> u16

Number of AlternateSet tables

pub fn alternate_set_offsets(&self) -> &'a [BigEndian<Offset16>]

Array of offsets to AlternateSet tables. Offsets are from beginning of substitution subtable, ordered by Coverage index

pub fn alternate_sets(&self) -> ArrayOfOffsets<'a, AlternateSet<'a>, Offset16>

A dynamically resolving wrapper for alternate_set_offsets.