read_fonts::tables::gsub

Type Alias LigatureSubstFormat1

pub type LigatureSubstFormat1<'a> = TableRef<'a, LigatureSubstFormat1Marker>;

Ligature Substitution Format 1

Aliased Type

struct LigatureSubstFormat1<'a> { /* private fields */ }

Implementations

impl<'a> LigatureSubstFormat1<'a>

pub fn subst_format(&self) -> u16

Format identifier: format = 1

pub fn coverage_offset(&self) -> Offset16

Offset to Coverage table, from beginning of substitution subtable

pub fn coverage(&self) -> Result<CoverageTable<'a>, ReadError>

Attempt to resolve coverage_offset.

pub fn ligature_set_count(&self) -> u16

Number of LigatureSet tables

pub fn ligature_set_offsets(&self) -> &'a [BigEndian<Offset16>]

Array of offsets to LigatureSet tables. Offsets are from beginning of substitution subtable, ordered by Coverage index

pub fn ligature_sets(&self) -> ArrayOfOffsets<'a, LigatureSet<'a>, Offset16>

A dynamically resolving wrapper for ligature_set_offsets.