read_fonts::tables::gsub

Type Alias ReverseChainSingleSubstFormat1

pub type ReverseChainSingleSubstFormat1<'a> = TableRef<'a, ReverseChainSingleSubstFormat1Marker>;

Reverse Chaining Contextual Single Substitution Format 1

Aliased Type

struct ReverseChainSingleSubstFormat1<'a> { /* private fields */ }

Implementations

impl<'a> ReverseChainSingleSubstFormat1<'a>

pub fn subst_format(&self) -> u16

Format identifier: format = 1

pub fn coverage_offset(&self) -> Offset16

Offset to Coverage table, from beginning of substitution subtable.

pub fn coverage(&self) -> Result<CoverageTable<'a>, ReadError>

Attempt to resolve coverage_offset.

pub fn backtrack_glyph_count(&self) -> u16

Number of glyphs in the backtrack sequence.

pub fn backtrack_coverage_offsets(&self) -> &'a [BigEndian<Offset16>]

Array of offsets to coverage tables in backtrack sequence, in glyph sequence order.

pub fn backtrack_coverages( &self, ) -> ArrayOfOffsets<'a, CoverageTable<'a>, Offset16>

A dynamically resolving wrapper for backtrack_coverage_offsets.

pub fn lookahead_glyph_count(&self) -> u16

Number of glyphs in lookahead sequence.

pub fn lookahead_coverage_offsets(&self) -> &'a [BigEndian<Offset16>]

Array of offsets to coverage tables in lookahead sequence, in glyph sequence order.

pub fn lookahead_coverages( &self, ) -> ArrayOfOffsets<'a, CoverageTable<'a>, Offset16>

A dynamically resolving wrapper for lookahead_coverage_offsets.

pub fn glyph_count(&self) -> u16

Number of glyph IDs in the substituteGlyphIDs array.

pub fn substitute_glyph_ids(&self) -> &'a [BigEndian<GlyphId16>]

Array of substitute glyph IDs — ordered by Coverage index.