read_fonts::tables

Module gvar

The gvar (Glyph Variations) table

Structs

GlyphDelta

Delta information for a single point or component in a glyph.

GvarFlags

U16Or32

Type Aliases

GlyphVariationData

Variation data specialized for the glyph variations table.

GlyphVariationDataHeader

The GlyphVariationData table

Gvar

The ‘gvar’ header

SharedTuples

Array of tuple records shared across all glyph variation data tables.