read_fonts::tables::gvar

Function find_glyph_and_point_count

fn find_glyph_and_point_count(
    glyf: &Glyf<'_>,
    loca: &Loca<'_>,
    glyph_id: GlyphId,
    recurse_depth: usize,
) -> Result<(GlyphId, usize), ReadError>

Given a glyph identifier, searches for the glyph that contains the actual metrics for rendering.

For simple glyphs, that is simply the requested glyph. For composites, it depends on the USE_MY_METRICS flag.

Returns the resulting glyph identifier and the number of points (or components) in that glyph. This count represents the start of the phantom points.