read_fonts::tables::hdmx

Type Alias Hdmx

pub type Hdmx<'a> = TableRef<'a, HdmxMarker>;

The Horizontal Device Metrics table.

Aliased Type

struct Hdmx<'a> { /* private fields */ }

Implementations

impl<'a> Hdmx<'a>

pub fn read(data: FontData<'a>, num_glyphs: u16) -> Result<Self, ReadError>

A constructor that requires additional arguments.

This type requires some external state in order to be parsed.

impl<'a> Hdmx<'a>

pub fn version(&self) -> u16

Table version number (set to 0).

pub fn num_records(&self) -> u16

Number of device records.

pub fn size_device_record(&self) -> u32

Size of device record, 32-bit aligned.

pub fn records(&self) -> ComputedArray<'a, DeviceRecord<'a>>

Array of device records.

impl<'a> Hdmx<'a>

pub fn record_for_size(&self, size: u8) -> Option<DeviceRecord<'a>>

Returns for the device record that exactly matches the given size (as ppem).