read_fonts::tables::head

Struct Flags

pub struct Flags { /* private fields */ }

The flags field for the head table.

Implementations

impl Flags

pub const BASELINE_AT_Y_0: Self

Bit 0: Baseline for font at y=0.

pub const LSB_AT_X_0: Self

Bit 1: Left sidebearing point at x=0 (relevant only for TrueType rasterizers).

pub const INSTRUCTIONS_MAY_DEPEND_ON_POINT_SIZE: Self

Bit 2: Instructions may depend on point size.

pub const FORCE_INTEGER_PPEM: Self

Bit 3: Force ppem to integer values for all internal scaler math; may use fractional ppem sizes if this bit is clear. It is strongly recommended that this be set in hinted fonts.

pub const INSTRUCTIONS_MAY_ALTER_ADVANCE_WIDTH: Self

Bit 4: Instructions may alter advance width (the advance widths might not scale linearly).

pub const LOSSLESS_TRANSFORMED_FONT_DATA: Self

Bit 11: Font data is “lossless” as a result of having been subjected to optimizing transformation and/or compression (such as compression mechanisms defined by ISO/IEC 14496-18, MicroType® Express, WOFF 2.0, or similar) where the original font functionality and features are retained but the binary compatibility between input and output font files is not guaranteed. As a result of the applied transform, the DSIG table may also be invalidated.

pub const CONVERTED_FONT: Self

Bit 12: Font converted (produce compatible metrics).

pub const OPTIMIZED_FOR_CLEARTYPE: Self

Bit 13: Font optimized for ClearType. Note, fonts that rely on embedded bitmaps (EBDT) for rendering should not be considered optimized for ClearType, and therefore should keep this bit cleared.

pub const LAST_RESORT_FONT: Self

Bit 14: Last Resort font. If set, indicates that the glyphs encoded in the ‘cmap’ subtables are simply generic symbolic representations of code point ranges and do not truly represent support for those code points. If unset, indicates that the glyphs encoded in the ‘cmap’ subtables represent proper support for those code points.

impl Flags

pub const fn empty() -> Self

Returns an empty set of flags.

pub const fn all() -> Self

Returns the set containing all flags.

pub const fn bits(&self) -> u16

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u16) -> Option<Self>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u16) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn intersects(&self, other: Self) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Self) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Self)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Self)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Self)

Toggles the specified flags in-place.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: Self) -> Self

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both.

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: Self) -> Self

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the “bit-clear” operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.