read_fonts::tables::hvar

Type Alias Hvar

pub type Hvar<'a> = TableRef<'a, HvarMarker>;

The HVAR (Horizontal Metrics Variations) table

Aliased Type

struct Hvar<'a> { /* private fields */ }

Implementations

impl<'a> Hvar<'a>

pub fn version(&self) -> MajorMinor

Major version number of the horizontal metrics variations table — set to 1. Minor version number of the horizontal metrics variations table — set to 0.

pub fn item_variation_store_offset(&self) -> Offset32

Offset in bytes from the start of this table to the item variation store table.

pub fn item_variation_store(&self) -> Result<ItemVariationStore<'a>, ReadError>

Attempt to resolve item_variation_store_offset.

pub fn advance_width_mapping_offset(&self) -> Nullable<Offset32>

Offset in bytes from the start of this table to the delta-set index mapping for advance widths (may be NULL).

pub fn advance_width_mapping( &self, ) -> Option<Result<DeltaSetIndexMap<'a>, ReadError>>

Attempt to resolve advance_width_mapping_offset.

pub fn lsb_mapping_offset(&self) -> Nullable<Offset32>

Offset in bytes from the start of this table to the delta-set index mapping for left side bearings (may be NULL).

pub fn lsb_mapping(&self) -> Option<Result<DeltaSetIndexMap<'a>, ReadError>>

Attempt to resolve lsb_mapping_offset.

pub fn rsb_mapping_offset(&self) -> Nullable<Offset32>

Offset in bytes from the start of this table to the delta-set index mapping for right side bearings (may be NULL).

pub fn rsb_mapping(&self) -> Option<Result<DeltaSetIndexMap<'a>, ReadError>>

Attempt to resolve rsb_mapping_offset.

impl Hvar<'_>

pub fn advance_width_delta( &self, glyph_id: GlyphId, coords: &[F2Dot14], ) -> Result<Fixed, ReadError>

Returns the advance width delta for the specified glyph identifier and normalized variation coordinates.

pub fn lsb_delta( &self, glyph_id: GlyphId, coords: &[F2Dot14], ) -> Result<Fixed, ReadError>

Returns the left side bearing delta for the specified glyph identifier and normalized variation coordinates.

pub fn rsb_delta( &self, glyph_id: GlyphId, coords: &[F2Dot14], ) -> Result<Fixed, ReadError>

Returns the left side bearing delta for the specified glyph identifier and normalized variation coordinates.