read_fonts::tables::ift

Type Alias GlyphPatches

pub type GlyphPatches<'a> = TableRef<'a, GlyphPatchesMarker>;

GlyphPatches

Aliased Type

struct GlyphPatches<'a> { /* private fields */ }

Implementations

impl<'a> GlyphPatches<'a>

pub fn read( data: FontData<'a>, flags: GlyphKeyedFlags, ) -> Result<Self, ReadError>

A constructor that requires additional arguments.

This type requires some external state in order to be parsed.

impl<'a> GlyphPatches<'a>

pub fn glyph_count(&self) -> u32

pub fn table_count(&self) -> u8

pub fn glyph_ids(&self) -> ComputedArray<'a, U16Or24>

pub fn tables(&self) -> &'a [BigEndian<Tag>]

pub fn glyph_data_offsets(&self) -> &'a [BigEndian<Offset32>]

pub fn glyph_data(&self) -> ArrayOfOffsets<'a, GlyphData<'a>, Offset32>

A dynamically resolving wrapper for glyph_data_offsets.

impl<'a> GlyphPatches<'a>

pub fn glyph_data_for_table( &'a self, table_index: usize, ) -> impl Iterator<Item = Result<(GlyphId, &'a [u8]), ReadError>>

Returns an iterator over the per glyph data for the table with the given index.