read_fonts::tables::ift

Type Alias PatchMapFormat1

pub type PatchMapFormat1<'a> = TableRef<'a, PatchMapFormat1Marker>;

Patch Map Format Format 1

Aliased Type

struct PatchMapFormat1<'a> { /* private fields */ }

Implementations

impl<'a> PatchMapFormat1<'a>

pub fn format(&self) -> u8

Format identifier: format = 1

pub fn compatibility_id(&self) -> CompatibilityId

Unique ID that identifies compatible patches.

pub fn max_entry_index(&self) -> u16

Largest entry index which appears in either the glyph map or feature map.

pub fn max_glyph_map_entry_index(&self) -> u16

Largest entry index which appears in the glyph map.

pub fn glyph_count(&self) -> Uint24

pub fn glyph_map_offset(&self) -> Offset32

Sub table that maps glyph ids to entry indices.

pub fn glyph_map(&self) -> Result<GlyphMap<'a>, ReadError>

Attempt to resolve glyph_map_offset.

pub fn feature_map_offset(&self) -> Nullable<Offset32>

Sub table that maps feature and glyph ids to entry indices.

pub fn feature_map(&self) -> Option<Result<FeatureMap<'a>, ReadError>>

Attempt to resolve feature_map_offset.

pub fn applied_entries_bitmap(&self) -> &'a [u8]

pub fn uri_template_length(&self) -> u16

pub fn uri_template(&self) -> &'a [u8]

pub fn patch_format(&self) -> u8

Patch format number for patches referenced by this mapping.

impl<'a> PatchMapFormat1<'a>

pub fn gid_to_entry_iter(&'a self) -> impl Iterator<Item = (GlyphId, u16)> + 'a

pub fn entry_count(&self) -> u32

pub fn uri_template_as_string(&self) -> Result<&str, ReadError>

pub fn is_entry_applied(&self, entry_index: u16) -> bool