read_fonts::tables::kerx

Type Alias Subtable0

pub type Subtable0<'a> = TableRef<'a, Subtable0Marker>;

The type 0 kerx subtable.

Aliased Type

pub struct Subtable0<'a> { /* private fields */ }

Implementations

impl<'a> Subtable0<'a>

pub fn n_pairs(&self) -> u32

The number of kerning pairs in this subtable.

pub fn search_range(&self) -> u32

The largest power of two less than or equal to the value of nPairs, multiplied by the size in bytes of an entry in the subtable.

pub fn entry_selector(&self) -> u32

This is calculated as log2 of the largest power of two less than or equal to the value of nPairs. This value indicates how many iterations of the search loop have to be made. For example, in a list of eight items, there would be three iterations of the loop.

pub fn range_shift(&self) -> u32

The value of nPairs minus the largest power of two less than or equal to nPairs. This is multiplied by the size in bytes of an entry in the table.

pub fn pairs(&self) -> &'a [Subtable0Pair]

Kerning records.

impl Subtable0<'_>

pub fn kerning(&self, left: GlyphId, right: GlyphId) -> Option<i32>

Returns the kerning adjustment for the given pair.