read_fonts::tables::layout

Enum Subtables

pub enum Subtables<'a, T: FontRead<'a>, Ext: ExtensionLookup<'a, T>> {
    Subtable(ArrayOfOffsets<'a, T>),
    Extension(ArrayOfOffsets<'a, Ext>),
}

an array of subtables, maybe behind extension lookups

This is used to implement more ergonomic access to lookup subtables for GPOS & GSUB lookup tables.

Variants

Subtable(ArrayOfOffsets<'a, T>)

Extension(ArrayOfOffsets<'a, Ext>)

Implementations

impl<'a, T: FontRead<'a> + 'a, Ext: ExtensionLookup<'a, T> + 'a> Subtables<'a, T, Ext>

pub fn len(&self) -> usize

The number of subtables in this collection

pub fn is_empty(&self) -> bool

pub fn get(&self, idx: usize) -> Result<T, ReadError>

Return the subtable at the given index

pub fn iter(&self) -> impl Iterator<Item = Result<T, ReadError>> + 'a

Return an iterator over all the subtables in the collection