read_fonts::tables::layout

Struct SizeParams

pub struct SizeParams<'a> { /* private fields */ }

Implementations

impl<'a> SizeParams<'a>

pub const MIN_SIZE: usize

pub fn resolve_offset<O: Offset, R: FontRead<'a>>( &self, offset: O, ) -> Result<R, ReadError>

Resolve the provided offset from the start of this table.

pub fn offset_data(&self) -> FontData<'a>

Return a reference to this table’s raw data.

We use this in the compile crate to resolve offsets.

pub fn shape(&self) -> &Self

👎Deprecated:
just use the base type directly

Return a reference to the table’s ‘Shape’ struct.

This is a low level implementation detail, but it can be useful in some cases where you want to know things about a table’s layout, such as the byte offsets of specific fields.

pub fn design_size(&self) -> u16

The first value represents the design size in 720/inch units (decipoints).

The design size entry must be non-zero. When there is a design size but no recommended size range, the rest of the array will consist of zeros.

pub fn identifier(&self) -> u16

The second value has no independent meaning, but serves as an identifier that associates fonts in a subfamily.

All fonts which share a Typographic or Font Family name and which differ only by size range shall have the same subfamily value, and no fonts which differ in weight or style shall have the same subfamily value. If this value is zero, the remaining fields in the array will be ignored.

pub fn name_entry(&self) -> u16

The third value enables applications to use a single name for the subfamily identified by the second value.

If the preceding value is non-zero, this value must be set in the range 256 – 32767 (inclusive). It records the value of a field in the ‘name’ table, which must contain English-language strings encoded in Windows Unicode and Macintosh Roman, and may contain additional strings localized to other scripts and languages. Each of these strings is the name an application should use, in combination with the family name, to represent the subfamily in a menu. Applications will choose the appropriate version based on their selection criteria.

pub fn range_start(&self) -> u16

The fourth and fifth values represent the small end of the recommended usage range (exclusive) and the large end of the recommended usage range (inclusive), stored in 720/inch units (decipoints).

Ranges must not overlap, and should generally be contiguous.

pub fn range_end(&self) -> u16

pub fn design_size_byte_range(&self) -> Range<usize>

pub fn identifier_byte_range(&self) -> Range<usize>

pub fn name_entry_byte_range(&self) -> Range<usize>

pub fn range_start_byte_range(&self) -> Range<usize>

pub fn range_end_byte_range(&self) -> Range<usize>