read_fonts::tables::layout

Trait ExtensionLookup

pub trait ExtensionLookup<'a, T: FontRead<'a>>: FontRead<'a> {
    // Required method
    fn extension(&self) -> Result<T, ReadError>;
}

A trait that abstracts the behaviour of an extension subtable

This is necessary because GPOS and GSUB have different concrete types for their extension lookups.

Required Methods

fn extension(&self) -> Result<T, ReadError>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.