Type Alias read_fonts::tables::layout::ChainedSequenceContextFormat2

pub type ChainedSequenceContextFormat2<'a> = TableRef<'a, ChainedSequenceContextFormat2Marker>;

Chained Sequence Context Format 2

Aliased Type

struct ChainedSequenceContextFormat2<'a> { /* private fields */ }

Implementations

impl<'a> ChainedSequenceContextFormat2<'a>

pub fn format(&self) -> u16

Format identifier: format = 2

pub fn coverage_offset(&self) -> Offset16

Offset to Coverage table, from beginning of ChainedSequenceContextFormat2 table

pub fn coverage(&self) -> Result<CoverageTable<'a>, ReadError>

Attempt to resolve coverage_offset.

pub fn backtrack_class_def_offset(&self) -> Offset16

Offset to ClassDef table containing backtrack sequence context, from beginning of ChainedSequenceContextFormat2 table

pub fn backtrack_class_def(&self) -> Result<ClassDef<'a>, ReadError>

Attempt to resolve backtrack_class_def_offset.

pub fn input_class_def_offset(&self) -> Offset16

Offset to ClassDef table containing input sequence context, from beginning of ChainedSequenceContextFormat2 table

pub fn input_class_def(&self) -> Result<ClassDef<'a>, ReadError>

Attempt to resolve input_class_def_offset.

pub fn lookahead_class_def_offset(&self) -> Offset16

Offset to ClassDef table containing lookahead sequence context, from beginning of ChainedSequenceContextFormat2 table

pub fn lookahead_class_def(&self) -> Result<ClassDef<'a>, ReadError>

Attempt to resolve lookahead_class_def_offset.

pub fn chained_class_seq_rule_set_count(&self) -> u16

Number of ChainedClassSequenceRuleSet tables

pub fn chained_class_seq_rule_set_offsets( &self ) -> &'a [BigEndian<Nullable<Offset16>>]

Array of offsets to ChainedClassSequenceRuleSet tables, from beginning of ChainedSequenceContextFormat2 table (may be NULL)

pub fn chained_class_seq_rule_sets( &self ) -> ArrayOfNullableOffsets<'a, ChainedClassSequenceRuleSet<'a>, Offset16>

A dynamically resolving wrapper for chained_class_seq_rule_set_offsets.