Type Alias read_fonts::tables::layout::CharacterVariantParams

pub type CharacterVariantParams<'a> = TableRef<'a, CharacterVariantParamsMarker>;

featureParams for ‘cv01’-‘cv99’

Aliased Type

struct CharacterVariantParams<'a> { /* private fields */ }

Implementations

impl<'a> CharacterVariantParams<'a>

pub fn format(&self) -> u16

Format number is set to 0.

pub fn feat_ui_label_name_id(&self) -> NameId

The ‘name’ table name ID that specifies a string (or strings, for multiple languages) for a user-interface label for this feature. (May be NULL.)

pub fn feat_ui_tooltip_text_name_id(&self) -> NameId

The ‘name’ table name ID that specifies a string (or strings, for multiple languages) that an application can use for tooltip text for this feature. (May be NULL.)

pub fn sample_text_name_id(&self) -> NameId

The ‘name’ table name ID that specifies sample text that illustrates the effect of this feature. (May be NULL.)

pub fn num_named_parameters(&self) -> u16

Number of named parameters. (May be zero.)

pub fn first_param_ui_label_name_id(&self) -> NameId

The first ‘name’ table name ID used to specify strings for user-interface labels for the feature parameters. (Must be zero if numParameters is zero.)

pub fn char_count(&self) -> u16

The count of characters for which this feature provides glyph variants. (May be zero.)

pub fn character(&self) -> &'a [BigEndian<Uint24>]

The Unicode Scalar Value of the characters for which this feature provides glyph variants.