Type Alias read_fonts::tables::layout::Feature

pub type Feature<'a> = TableRef<'a, FeatureMarker>;

Feature Table

Aliased Type

struct Feature<'a> { /* private fields */ }

Implementations

impl<'a> Feature<'a>

pub fn read(data: FontData<'a>, feature_tag: Tag) -> Result<Self, ReadError>

A constructor that requires additional arguments.

This type requires some external state in order to be parsed.

impl<'a> Feature<'a>

pub fn feature_params_offset(&self) -> Nullable<Offset16>

Offset from start of Feature table to FeatureParams table, if defined for the feature and present, else NULL

pub fn feature_params(&self) -> Option<Result<FeatureParams<'a>, ReadError>>

Attempt to resolve feature_params_offset.

pub fn lookup_index_count(&self) -> u16

Number of LookupList indices for this feature

pub fn lookup_list_indices(&self) -> &'a [BigEndian<u16>]

Array of indices into the LookupList — zero-based (first lookup is LookupListIndex = 0)