read_fonts::tables::layout

Type Alias LangSys

pub type LangSys<'a> = TableRef<'a, LangSysMarker>;

Language System Table

Aliased Type

struct LangSys<'a> { /* private fields */ }

Implementations

impl LangSys<'_>

pub fn feature_index_for_tag( &self, list: &FeatureList<'_>, tag: Tag, ) -> Option<u16>

If the language system references a feature with the given tag, returns the index of that feature in the specified feature list.

The feature list can be obtained from the feature_list method on the parent Gsub or Gpos tables.

impl<'a> LangSys<'a>

pub fn required_feature_index(&self) -> u16

Index of a feature required for this language system; if no required features = 0xFFFF

pub fn feature_index_count(&self) -> u16

Number of feature index values for this language system — excludes the required feature

pub fn feature_indices(&self) -> &'a [BigEndian<u16>]

Array of indices into the FeatureList, in arbitrary order