read_fonts::tables::layout

Type Alias LookupList

pub type LookupList<'a, T> = TableRef<'a, LookupListMarker<T>>;

Lookup List Table

Aliased Type

pub struct LookupList<'a, T> {
    pub(crate) shape: LookupListMarker<T>,
    pub(crate) data: FontData<'a>,
}

Fields
shape: LookupListMarker<T>data: FontData<'a>
Implementations

impl<'a> LookupList<'a, ()>

pub(crate) fn into_concrete<T>(self) -> LookupList<'a, T>

impl<'a, T> LookupList<'a, T>

pub(crate) fn of_unit_type(&self) -> LookupList<'a, ()>

Replace the specific generic type on this implementation with ()

impl<'a, T> LookupList<'a, T>

pub fn lookup_count(&self) -> u16

Number of lookups in this table

pub fn lookup_offsets(&self) -> &'a [BigEndian<Offset16>]

Array of offsets to Lookup tables, from beginning of LookupList — zero based (first lookup is Lookup index = 0)

pub fn lookups(&self) -> ArrayOfOffsets<'a, T, Offset16>
where T: FontRead<'a>,

A dynamically resolving wrapper for lookup_offsets.