read_fonts::tables::layout

Type Alias ScriptList

pub type ScriptList<'a> = TableRef<'a, ScriptListMarker>;

Script List Table

Aliased Type

struct ScriptList<'a> { /* private fields */ }

Implementations

impl<'a> ScriptList<'a>

pub fn index_for_tag(&self, tag: Tag) -> Option<u16>

Returns the index of the script with the given tag.

pub fn get(&self, index: u16) -> Result<TaggedElement<Script<'a>>, ReadError>

Returns the tag and script at the given index.

pub fn select(&self, tags: &[Tag]) -> Option<SelectedScript>

Finds the first available script that matches one of the given tags.

When none of the requested scripts are available, then DFLT, dflt and latn tags are tried in that order.

If you’re starting from a Unicode script code, use the ScriptTags::from_unicode function to generate the appropriate set of tags to pass to this method.

See hb_ot_layout_table_select_script for the corresponding HarfBuzz function.

impl<'a> ScriptList<'a>

pub fn script_count(&self) -> u16

Number of ScriptRecords

pub fn script_records(&self) -> &'a [ScriptRecord]

Array of ScriptRecords, listed alphabetically by script tag