read_fonts::tables::layout

Type Alias StylisticSetParams

pub type StylisticSetParams<'a> = TableRef<'a, StylisticSetParamsMarker>;

Aliased Type

struct StylisticSetParams<'a> { /* private fields */ }

Implementations

impl<'a> StylisticSetParams<'a>

pub fn version(&self) -> u16

pub fn ui_name_id(&self) -> NameId

The ‘name’ table name ID that specifies a string (or strings, for multiple languages) for a user-interface label for this feature.

The value of uiLabelNameId is expected to be in the font-specific name ID range (256-32767), though that is not a requirement in this Feature Parameters specification. The user-interface label for the feature can be provided in multiple languages. An English string should be included as a fallback. The string should be kept to a minimal length to fit comfortably with different application interfaces.