read_fonts::tables::morx

Type Alias Subtable

pub type Subtable<'a> = TableRef<'a, SubtableMarker>;

A subtable in a morx chain.

Aliased Type

pub struct Subtable<'a> { /* private fields */ }

Implementations

impl<'a> Subtable<'a>

pub fn length(&self) -> u32

Total subtable length, including this header.

pub fn coverage(&self) -> u32

Coverage flags and subtable type.

pub fn sub_feature_flags(&self) -> u32

The 32-bit mask identifying which subtable this is (the subtable being executed if the AND of this value and the processed defaultFlags is nonzero).

pub fn data(&self) -> &'a [u8]

Data for specific subtable.

impl<'a> Subtable<'a>

pub fn is_logical(&self) -> bool

If true, this subtable will process glyphs in logical order (or reverse logical order, depending on the value of bit 0x80000000).

pub fn is_all_directions(&self) -> bool

If true, this subtable will be applied to both horizontal and vertical text (i.e. the state of bit 0x80000000 is ignored).

pub fn is_backwards(&self) -> bool

If true, this subtable will process glyphs in descending order. Otherwise, it will process the glyphs in ascending order.

pub fn is_vertical(&self) -> bool

If true, this subtable will only be applied to vertical text. Otherwise, this subtable will only be applied to horizontal text.

pub fn kind(&self) -> Result<SubtableKind<'a>, ReadError>

Returns an enum representing the actual subtable data.