read_fonts::tables::mvar

Type Alias Mvar

pub type Mvar<'a> = TableRef<'a, MvarMarker>;

The MVAR (Metrics Variations) table

Aliased Type

struct Mvar<'a> { /* private fields */ }

Implementations

impl<'a> Mvar<'a>

pub fn version(&self) -> MajorMinor

Major version number of the horizontal metrics variations table — set to 1. Minor version number of the horizontal metrics variations table — set to 0.

pub fn value_record_size(&self) -> u16

The size in bytes of each value record — must be greater than zero.

pub fn value_record_count(&self) -> u16

The number of value records — may be zero.

pub fn item_variation_store_offset(&self) -> Nullable<Offset16>

Offset in bytes from the start of this table to the item variation store table. If valueRecordCount is zero, set to zero; if valueRecordCount is greater than zero, must be greater than zero.

pub fn item_variation_store( &self, ) -> Option<Result<ItemVariationStore<'a>, ReadError>>

Attempt to resolve item_variation_store_offset.

pub fn value_records(&self) -> &'a [ValueRecord]

Array of value records that identify target items and the associated delta-set index for each. The valueTag records must be in binary order of their valueTag field.

impl Mvar<'_>

pub fn metric_delta( &self, tag: Tag, coords: &[F2Dot14], ) -> Result<Fixed, ReadError>

Returns the metric delta for the specified tag and normalized variation coordinates. Possible tags are found in the tags module.