read_fonts::tables::post

Type Alias Post

pub type Post<'a> = TableRef<'a, PostMarker>;

post (PostScript) table

Aliased Type

struct Post<'a> { /* private fields */ }

Implementations

impl<'a> Post<'a>

pub fn version(&self) -> Version16Dot16

0x00010000 for version 1.0 0x00020000 for version 2.0 0x00025000 for version 2.5 (deprecated) 0x00030000 for version 3.0

pub fn italic_angle(&self) -> Fixed

Italic angle in counter-clockwise degrees from the vertical. Zero for upright text, negative for text that leans to the right (forward).

pub fn underline_position(&self) -> FWord

This is the suggested distance of the top of the underline from the baseline (negative values indicate below baseline). The PostScript definition of this FontInfo dictionary key (the y coordinate of the center of the stroke) is not used for historical reasons. The value of the PostScript key may be calculated by subtracting half the underlineThickness from the value of this field.

pub fn underline_thickness(&self) -> FWord

Suggested values for the underline thickness. In general, the underline thickness should match the thickness of the underscore character (U+005F LOW LINE), and should also match the strikeout thickness, which is specified in the OS/2 table.

pub fn is_fixed_pitch(&self) -> u32

Set to 0 if the font is proportionally spaced, non-zero if the font is not proportionally spaced (i.e. monospaced).

pub fn min_mem_type42(&self) -> u32

Minimum memory usage when an OpenType font is downloaded.

pub fn max_mem_type42(&self) -> u32

Maximum memory usage when an OpenType font is downloaded.

pub fn min_mem_type1(&self) -> u32

Minimum memory usage when an OpenType font is downloaded as a Type 1 font.

pub fn max_mem_type1(&self) -> u32

Maximum memory usage when an OpenType font is downloaded as a Type 1 font.

pub fn num_glyphs(&self) -> Option<u16>

Number of glyphs (this should be the same as numGlyphs in ‘maxp’ table).

pub fn glyph_name_index(&self) -> Option<&'a [BigEndian<u16>]>

Array of indices into the string data. See below for details.

pub fn string_data(&self) -> Option<VarLenArray<'a, PString<'a>>>

Storage for the string data.

impl<'a> Post<'a>

pub fn num_names(&self) -> usize

The number of glyph names covered by this table

pub fn glyph_name(&self, glyph_id: GlyphId16) -> Option<&str>