read_fonts::tables

Module postscript

PostScript (CFF and CFF2) common tables.

Modules

charstring

Parsing for PostScript charstrings.

dict

Parsing for PostScript DICTs.

Structs

BlendState

State for processing the blend operator for DICTs and charstrings.

Charset

Character set for mapping from glyph to string identifiers.

CharsetIter

Iterator over the glyph and string identifier mappings in a character set.

CharsetRange1

Range struct for Charset format 1.

CharsetRange2

Range struct for Charset format 2.

FdSelectRange3

Range struct for FdSelect format 3.

FdSelectRange4

Range struct for FdSelect format 4.

Latin1String

Reference to a Latin-1 encoded string.

Stack

Operand stack for DICTs and charstrings.

StringId

PostScript string identifier (SID).

Enums

CustomCharset

Charset with custom glyph id to string id mappings.

Error

Errors that are specific to PostScript processing.

FdSelect

Associates a glyph identifier with a Font DICT.

Index

Common type for uniform access to CFF and CFF2 index formats.

Number

Either a signed 32-bit integer or a 16.16 fixed point number.

Constants

STANDARD_STRINGS

The PostScript standard string set.

Type Aliases

CharsetFormat0

Charset format 0.

CharsetFormat1

Charset format 1.

CharsetFormat2

Charset format 2.

FdSelectFormat0

FdSelect format 0.

FdSelectFormat3

FdSelect format 3.

FdSelectFormat4

FdSelect format 4.

Index1

An array of variable-sized objects in a CFF table.

Index2

An array of variable-sized objects in a CFF2 table.