Function read_fonts::tables::postscript::dict::entries

pub fn entries<'a>(
    dict_data: &'a [u8],
    blend_state: Option<BlendState<'a>>
) -> impl Iterator<Item = Result<Entry, Error>> + 'a

Given a byte slice containing DICT data, returns an iterator yielding each operator with its associated operands.

This performs appropriate type conversions, decodes deltas and applies blending.

If processing a Private DICT from a CFF2 table and an item variation store is present, then blend_state must be provided.