read_fonts::tables::postscript

Type Alias Index1

pub type Index1<'a> = TableRef<'a, Index1Marker>;

An array of variable-sized objects in a CFF table.

Aliased Type

struct Index1<'a> { /* private fields */ }

Implementations

impl<'a> Index1<'a>

pub fn size_in_bytes(&self) -> Result<usize, ReadError>

Returns the total size in bytes of the index table.

pub fn get_offset(&self, index: usize) -> Result<usize, Error>

Returns the offset of the object at the given index.

pub fn get(&self, index: usize) -> Result<&'a [u8], Error>

Returns the data for the object at the given index.

impl<'a> Index1<'a>

pub fn count(&self) -> u16

Number of objects stored in INDEX.

pub fn off_size(&self) -> u8

Object array element size.

pub fn offsets(&self) -> &'a [u8]

Bytes containing count + 1 offsets each of off_size.

pub fn data(&self) -> &'a [u8]

Array containing the object data.