Type Alias read_fonts::tables::sbix::GlyphData

pub type GlyphData<'a> = TableRef<'a, GlyphDataMarker>;

Glyph data table

Aliased Type

struct GlyphData<'a> { /* private fields */ }

Implementations

impl<'a> GlyphData<'a>

pub fn origin_offset_x(&self) -> i16

The horizontal (x-axis) position of the left edge of the bitmap graphic in relation to the glyph design space origin.

pub fn origin_offset_y(&self) -> i16

The vertical (y-axis) position of the bottom edge of the bitmap graphic in relation to the glyph design space origin.

pub fn graphic_type(&self) -> Tag

Indicates the format of the embedded graphic data: one of ’jpg ’, ’png ’ or ‘tiff’, or the special format ‘dupe’.

pub fn data(&self) -> &'a [u8]

The actual embedded graphic data. The total length is inferred from sequential entries in the glyphDataOffsets array and the fixed size (8 bytes) of the preceding fields.