read_fonts::tables::sbix

Type Alias Strike

pub type Strike<'a> = TableRef<'a, StrikeMarker>;

Strike header table

Aliased Type

struct Strike<'a> { /* private fields */ }

Implementations

impl<'a> Strike<'a>

pub fn read(data: FontData<'a>, num_glyphs: u16) -> Result<Self, ReadError>

A constructor that requires additional arguments.

This type requires some external state in order to be parsed.

impl<'a> Strike<'a>

pub fn ppem(&self) -> u16

The PPEM size for which this strike was designed.

pub fn ppi(&self) -> u16

The device pixel density (in PPI) for which this strike was designed. (E.g., 96 PPI, 192 PPI.)

pub fn glyph_data_offsets(&self) -> &'a [BigEndian<u32>]

Offset from the beginning of the strike data header to bitmap data for an individual glyph ID.

impl<'a> Strike<'a>

pub fn glyph_data( &self, glyph_id: GlyphId, ) -> Result<Option<GlyphData<'a>>, ReadError>