read_fonts::tables::stat

Type Alias AxisValueFormat1

pub type AxisValueFormat1<'a> = TableRef<'a, AxisValueFormat1Marker>;

Axis value table format 1

Aliased Type

pub struct AxisValueFormat1<'a> {
    pub(crate) shape: AxisValueFormat1Marker,
    pub(crate) data: FontData<'a>,
}

Fields
shape: AxisValueFormat1Markerdata: FontData<'a>
Implementations

impl<'a> AxisValueFormat1<'a>

pub fn format(&self) -> u16

Format identifier — set to 1.

pub fn axis_index(&self) -> u16

Zero-base index into the axis record array identifying the axis of design variation to which the axis value table applies. Must be less than designAxisCount.

pub fn flags(&self) -> AxisValueTableFlags

Flags — see below for details.

pub fn value_name_id(&self) -> NameId

The name ID for entries in the ‘name’ table that provide a display string for this attribute value.

pub fn value(&self) -> Fixed

A numeric value for this attribute value.