Type Alias read_fonts::tables::stat::AxisValueFormat3

pub type AxisValueFormat3<'a> = TableRef<'a, AxisValueFormat3Marker>;

Axis value table format 3

Aliased Type

struct AxisValueFormat3<'a> { /* private fields */ }

Implementations

impl<'a> AxisValueFormat3<'a>

pub fn format(&self) -> u16

Format identifier — set to 3.

pub fn axis_index(&self) -> u16

Zero-base index into the axis record array identifying the axis of design variation to which the axis value table applies. Must be less than designAxisCount.

pub fn flags(&self) -> AxisValueTableFlags

Flags — see below for details.

pub fn value_name_id(&self) -> NameId

The name ID for entries in the ‘name’ table that provide a display string for this attribute value.

pub fn value(&self) -> Fixed

A numeric value for this attribute value.

pub fn linked_value(&self) -> Fixed

The numeric value for a style-linked mapping from this value.