read_fonts::tables::variations

Function compute_scalar

fn compute_scalar<'a>(
    header: &TupleVariationHeader<'_>,
    axis_count: usize,
    shared_tuples: &Option<ComputedArray<'a, Tuple<'a>>>,
    coords: &[F2Dot14],
) -> Option<Fixed>

Compute the fixed point scalar for this tuple at the given location in variation space.

The coords slice must be of lesser or equal length to the number of axes. If it is less, missing (trailing) axes will be assumed to have zero values.

Returns None if this tuple is not applicable at the provided coordinates (e.g. if the resulting scalar is zero).