Trait read_fonts::tables::variations::TupleDelta

pub trait TupleDelta: Sized + Copy {
    // Required methods
    fn is_point() -> bool;
    fn new(position: u16, x: i32, y: i32) -> Self;
}

Trait for deltas that are computed in a tuple variation store.

Required Methods

fn is_point() -> bool

Returns true if the delta is a point and requires reading two values from the packed delta stream.

fn new(position: u16, x: i32, y: i32) -> Self

Creates a new delta for the given position and coordinates. If the delta is not a point, the y value will always be zero.

Object Safety

This trait is not object safe.