read_fonts::tables::variations

Type Alias ItemVariationData

pub type ItemVariationData<'a> = TableRef<'a, ItemVariationDataMarker>;

The ItemVariationData subtable

Aliased Type

pub struct ItemVariationData<'a> { /* private fields */ }

Implementations

impl<'a> ItemVariationData<'a>

pub fn item_count(&self) -> u16

The number of delta sets for distinct items.

pub fn word_delta_count(&self) -> u16

A packed field: the high bit is a flag—see details below.

pub fn region_index_count(&self) -> u16

The number of variation regions referenced.

pub fn region_indexes(&self) -> &'a [BigEndian<u16>]

Array of indices into the variation region list for the regions referenced by this item variation data table.

pub fn delta_sets(&self) -> &'a [u8]

Delta-set rows.

impl<'a> ItemVariationData<'a>

pub fn delta_set( &self, inner_index: u16, ) -> impl Iterator<Item = i32> + 'a + Clone

Returns an iterator over the per-region delta values for the specified inner index.

pub fn get_delta_row_len(&self) -> usize

pub fn delta_row_len(word_delta_count: u16, region_index_count: u16) -> usize

the length of one delta set

pub fn delta_sets_len( item_count: u16, word_delta_count: u16, region_index_count: u16, ) -> usize