Type Alias read_fonts::tables::variations::ItemVariationStore

pub type ItemVariationStore<'a> = TableRef<'a, ItemVariationStoreMarker>;

The ItemVariationStore table

Aliased Type

struct ItemVariationStore<'a> { /* private fields */ }

Implementations

impl<'a> ItemVariationStore<'a>

pub fn format(&self) -> u16

Format— set to 1

pub fn variation_region_list_offset(&self) -> Offset32

Offset in bytes from the start of the item variation store to the variation region list.

pub fn variation_region_list( &self ) -> Result<VariationRegionList<'a>, ReadError>

Attempt to resolve variation_region_list_offset.

pub fn item_variation_data_count(&self) -> u16

The number of item variation data subtables.

pub fn item_variation_data_offsets(&self) -> &'a [BigEndian<Nullable<Offset32>>]

Offsets in bytes from the start of the item variation store to each item variation data subtable.

pub fn item_variation_data( &self ) -> ArrayOfNullableOffsets<'a, ItemVariationData<'a>, Offset32>

A dynamically resolving wrapper for item_variation_data_offsets.

impl<'a> ItemVariationStore<'a>

pub fn compute_delta( &self, index: DeltaSetIndex, coords: &[F2Dot14] ) -> Result<i32, ReadError>

Computes the delta value for the specified index and set of normalized variation coordinates.

pub fn compute_float_delta( &self, index: DeltaSetIndex, coords: &[F2Dot14] ) -> Result<FloatItemDelta, ReadError>

Computes the delta value in floating point for the specified index and set of normalized variation coordinates.