read_fonts::tables::variations

Type Alias TupleVariationHeader

pub type TupleVariationHeader<'a> = TableRef<'a, TupleVariationHeaderMarker>;

TupleVariationHeader

Aliased Type

struct TupleVariationHeader<'a> { /* private fields */ }

Implementations

impl<'a> TupleVariationHeader<'a>

pub fn read(data: FontData<'a>, axis_count: u16) -> Result<Self, ReadError>

A constructor that requires additional arguments.

This type requires some external state in order to be parsed.

impl<'a> TupleVariationHeader<'a>

pub fn variation_data_size(&self) -> u16

The size in bytes of the serialized data for this tuple variation table.

pub fn tuple_index(&self) -> TupleIndex

A packed field. The high 4 bits are flags (see below). The low 12 bits are an index into a shared tuple records array.

impl<'a> TupleVariationHeader<'a>

pub fn peak_tuple(&self) -> Option<Tuple<'a>>

Peak tuple record for this tuple variation table — optional, determined by flags in the tupleIndex value. Note that this must always be included in the ‘cvar’ table.

pub fn intermediate_start_tuple(&self) -> Option<Tuple<'a>>

Intermediate start tuple record for this tuple variation table — optional, determined by flags in the tupleIndex value.

pub fn intermediate_end_tuple(&self) -> Option<Tuple<'a>>

Intermediate end tuple record for this tuple variation table — optional, determined by flags in the tupleIndex value.