read_fonts::tables::vmtx

Type Alias Vmtx

pub type Vmtx<'a> = TableRef<'a, VmtxMarker>;

The vmtx (Vertical Metrics) table

Aliased Type

struct Vmtx<'a> { /* private fields */ }

Implementations

impl<'a> Vmtx<'a>

pub fn read( data: FontData<'a>, number_of_long_ver_metrics: u16, num_glyphs: u16, ) -> Result<Self, ReadError>

A constructor that requires additional arguments.

This type requires some external state in order to be parsed.

impl<'a> Vmtx<'a>

pub fn v_metrics(&self) -> &'a [LongMetric]

Paired advance height and top side bearing values for each glyph. Records are indexed by glyph ID.

pub fn top_side_bearings(&self) -> &'a [BigEndian<i16>]

Top side bearings for glyph IDs greater than or equal to numberOfLongMetrics.

impl Vmtx<'_>

pub fn advance(&self, glyph_id: GlyphId) -> Option<u16>

Returns the advance height for the given glyph identifier.

pub fn side_bearing(&self, glyph_id: GlyphId) -> Option<i16>

Returns the top side bearing for the given glyph identifier.