read_fonts::tables::vorg

Type Alias Vorg

pub type Vorg<'a> = TableRef<'a, VorgMarker>;

The VORG (Vertical Origin) table.

Aliased Type

struct Vorg<'a> { /* private fields */ }

Implementations

impl<'a> Vorg<'a>

pub fn version(&self) -> MajorMinor

Major/minor version number. Set to 1.0.

pub fn default_vert_origin_y(&self) -> i16

The y coordinate of a glyph’s vertical origin, in the font’s design coordinate system, to be used if no entry is present for the glyph in the vertOriginYMetrics array.

pub fn num_vert_origin_y_metrics(&self) -> u16

Number of elements in the vertOriginYMetrics array.

pub fn vert_origin_y_metrics(&self) -> &'a [VertOriginYMetrics]

Array of VertOriginYMetrics records, sorted by glyph ID.

impl Vorg<'_>

pub fn vertical_origin_y(&self, glyph_id: GlyphId) -> i16

Returns the y coordinate of the of the glyph’s vertical origin.