Type Alias read_fonts::tables::vvar::Vvar

pub type Vvar<'a> = TableRef<'a, VvarMarker>;

The VVAR (Vertical Metrics Variations) table

Aliased Type

struct Vvar<'a> { /* private fields */ }

Implementations

impl<'a> Vvar<'a>

pub fn version(&self) -> MajorMinor

Major version number of the horizontal metrics variations table — set to 1. Minor version number of the horizontal metrics variations table — set to 0.

pub fn item_variation_store_offset(&self) -> Offset32

Offset in bytes from the start of this table to the item variation store table.

pub fn item_variation_store(&self) -> Result<ItemVariationStore<'a>, ReadError>

Attempt to resolve item_variation_store_offset.

pub fn advance_height_mapping_offset(&self) -> Nullable<Offset32>

Offset in bytes from the start of this table to the delta-set index mapping for advance heights (may be NULL).

pub fn advance_height_mapping( &self, ) -> Option<Result<DeltaSetIndexMap<'a>, ReadError>>

Attempt to resolve advance_height_mapping_offset.

pub fn tsb_mapping_offset(&self) -> Nullable<Offset32>

Offset in bytes from the start of this table to the delta-set index mapping for top side bearings (may be NULL).

pub fn tsb_mapping(&self) -> Option<Result<DeltaSetIndexMap<'a>, ReadError>>

Attempt to resolve tsb_mapping_offset.

pub fn bsb_mapping_offset(&self) -> Nullable<Offset32>

Offset in bytes from the start of this table to the delta-set index mapping for bottom side bearings (may be NULL).

pub fn bsb_mapping(&self) -> Option<Result<DeltaSetIndexMap<'a>, ReadError>>

Attempt to resolve bsb_mapping_offset.

pub fn v_org_mapping_offset(&self) -> Nullable<Offset32>

Offset in bytes from the start of this table to the delta-set index mapping for Y coordinates of vertical origins (may be NULL).

pub fn v_org_mapping(&self) -> Option<Result<DeltaSetIndexMap<'a>, ReadError>>

Attempt to resolve v_org_mapping_offset.

impl<'a> Vvar<'a>

pub fn advance_height_delta( &self, glyph_id: GlyphId, coords: &[F2Dot14], ) -> Result<Fixed, ReadError>

Returns the advance height delta for the specified glyph identifier and normalized variation coordinates.

pub fn tsb_delta( &self, glyph_id: GlyphId, coords: &[F2Dot14], ) -> Result<Fixed, ReadError>

Returns the top side bearing delta for the specified glyph identifier and normalized variation coordinates.

pub fn bsb_delta( &self, glyph_id: GlyphId, coords: &[F2Dot14], ) -> Result<Fixed, ReadError>

Returns the bottom side bearing delta for the specified glyph identifier and normalized variation coordinates.

pub fn v_org_delta( &self, glyph_id: GlyphId, coords: &[F2Dot14], ) -> Result<Fixed, ReadError>

Returns the vertical origin delta for the specified glyph identifier and normalized variation coordinates.