read_fonts

Trait FontRead

pub trait FontRead<'a>: Sized {
    // Required method
    fn read(data: FontData<'a>) -> Result<Self, ReadError>;
}

A type that can be read from raw table data.

This trait is implemented for all font tables that are self-describing: that is, tables that do not require any external state in order to interpret their underlying bytes. (Tables that require external state implement FontReadWithArgs instead)

Required Methods

fn read(data: FontData<'a>) -> Result<Self, ReadError>

Read an instance of Self from the provided data, performing validation.

In the case of a table, this method is responsible for ensuring the input data is consistent: this means ensuring that any versioned fields are present as required by the version, and that any array lengths are not out-of-bounds.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.