Trait read_fonts::FromBytes

pub unsafe trait FromBytes: FixedSize + Sealed { }

A marker trait for types that can read from a big-endian buffer without copying.

This is used as a trait bound on certain methods on FontData (such as FontData::read_ref_at and FontData::read_array) in order to ensure that those methods are only used with types that uphold certain safety guarantees.

WARNING: Do not implement this trait manually. Implementations are created where appropriate during code generation, and there should be no conditions under which this trait could be implemented, but cannot be implemented by codegen.

Safety

If a type T implements FromBytes then unsafe code may assume that it is safe to interpret any sequence of bytes with length equal to std::mem::size_of::<T>() as T.

we additionally ensure the following conditions:

the type must have no internal padding

std::mem::align_of::<T>() == 1

for structs, the type is repr(packed) and repr(C), and all fields are also FromBytes

In practice, this trait is only implemented for u8, BigEndian<T>, and for structs where all fields are those base types.

Object Safety

This trait is not object safe.