Trait read_fonts::traversal::SomeArray

pub trait SomeArray<'a> {
    // Required methods
    fn type_name(&self) -> &str;
    fn len(&self) -> usize;
    fn get(&self, idx: usize) -> Option<FieldType<'a>>;

    // Provided method
    fn is_empty(&self) -> bool { ... }
}

A generic trait for arrays.

Required Methods

fn type_name(&self) -> &str

The name of this type. For an array of u16s, this is [u16].

fn len(&self) -> usize

The length of the array.

fn get(&self, idx: usize) -> Option<FieldType<'a>>

Return the item at idx, or None if idx is out of bounds.

Provided Methods

fn is_empty(&self) -> bool

Returns true if this array is empty.

Implementations

impl<'a> dyn SomeArray<'a> + 'a

pub fn iter(&self) -> impl Iterator<Item = FieldType<'a>> + '_

Return an iterator over the contents of this array.