read_fonts

Module read

Traits for interpreting font data

Enums

ReadError

An error that occurs when reading font data

Traits

ComputeSize

A type that can compute its size at runtime, based on some input.

FontRead

A type that can be read from raw table data.

FontReadWithArgs

A trait for types that require external data in order to be constructed.

Format

A trait for tables that have multiple possible formats.

ReadArgs

A trait for a type that needs additional arguments to be read.

VarSize

A trait for types that have variable length.