read_fonts::tables

Module aat

Apple Advanced Typography common tables.

See https://developer.apple.com/fonts/TrueType-Reference-Manual/RM06/Chap6Tables.html

Modules

class

Predefined classes.

Structs

ExtendedStateTable

LookupSegment2

Lookup segment for format 2.

LookupSegment4

Lookup segment for format 4.

LookupSingle

Lookup single record for format 6.

NoPayload

Empty data type for a state table entry with no payload.

StateEntry

Entry in an (extended) state table.

StateTable

TypedLookup

Enums

Lookup

Lookup tables provide a way of looking up information about a glyph index. The different cmap subtable formats.

Traits

LookupValue

Trait for values that can be read from lookup tables.

Type Aliases

ClassSubtable

Maps the glyph indexes of your font into classes.

ExtendedStateTableU16

Lookup0

Simple array format. The lookup data is an array of lookup values, indexed by glyph index.

Lookup2

Segment single format. Each non-overlapping segment has a single lookup value that applies to all glyphs in the segment. A segment is defined as a contiguous range of glyph indexes.

Lookup4

Segment array format. A segment mapping is performed (as with Format 2), but instead of a single lookup value for all the glyphs in the segment, each glyph in the segment gets its own separate lookup value.

Lookup6

Single table format. The lookup data is a sorted list of <glyph index,lookup value> pairs.

Lookup8

Trimmed array format. The lookup data is a simple trimmed array indexed by glyph index.

Lookup10

Trimmed array format. The lookup data is a simple trimmed array indexed by glyph index.

LookupGlyphId

LookupU16

LookupU32

RawBytes

Used for the state_array and entry_table fields in StateHeader.

RawWords

Used for the state_array in StxHeader.

StateHeader

Header for a state table.

StxHeader

Header for an extended state table.