read_fonts::tables::aat

Type Alias Lookup4

pub type Lookup4<'a> = TableRef<'a, Lookup4Marker>;

Segment array format. A segment mapping is performed (as with Format 2), but instead of a single lookup value for all the glyphs in the segment, each glyph in the segment gets its own separate lookup value.

Aliased Type

struct Lookup4<'a> { /* private fields */ }

Implementations

impl<'a> Lookup4<'a>

pub fn format(&self) -> u16

Format number is set to 4.

pub fn unit_size(&self) -> u16

Size of a lookup unit for this search in bytes.

pub fn n_units(&self) -> u16

Number of units of the preceding size to be searched.

pub fn search_range(&self) -> u16

The value of unitSize times the largest power of 2 that is less than or equal to the value of nUnits.

pub fn entry_selector(&self) -> u16

The log base 2 of the largest power of 2 less than or equal to the value of nUnits.

pub fn range_shift(&self) -> u16

The value of unitSize times the difference of the value of nUnits minus the largest power of 2 less than or equal to the value of nUnits.

pub fn segments(&self) -> &'a [LookupSegment4]

Segments.

impl Lookup4<'_>

pub fn value<T: LookupValue>(&self, index: u16) -> Result<T, ReadError>