read_fonts::tables::cmap

Type Alias Cmap

pub type Cmap<'a> = TableRef<'a, CmapMarker>;

cmap

Aliased Type

struct Cmap<'a> { /* private fields */ }

Implementations

impl<'a> Cmap<'a>

pub fn version(&self) -> u16

Table version number (0).

pub fn num_tables(&self) -> u16

Number of encoding tables that follow.

pub fn encoding_records(&self) -> &'a [EncodingRecord]

impl Cmap<'_>

pub fn map_codepoint(&self, codepoint: impl Into<u32>) -> Option<GlyphId>

Map a codepoint to a nominal glyph identifier

This uses the first available subtable that provides a valid mapping.

Note:

Mapping logic is currently only implemented for the most common subtable formats.

pub fn closure_glyphs( &self, unicodes: &IntSet<u32>, glyph_set: &mut IntSet<GlyphId>, )