read_fonts::tables::cmap

Type Alias Cmap4

pub type Cmap4<'a> = TableRef<'a, Cmap4Marker>;

cmap Format 4: Segment mapping to delta values

Aliased Type

pub struct Cmap4<'a> {
    pub(crate) shape: Cmap4Marker,
    pub(crate) data: FontData<'a>,
}

Fields
shape: Cmap4Markerdata: FontData<'a>
Implementations

impl<'a> Cmap4<'a>

pub fn format(&self) -> u16

Format number is set to 4.

pub fn length(&self) -> u16

This is the length in bytes of the subtable.

pub fn language(&self) -> u16

For requirements on use of the language field, see “Use of the language field in ‘cmap’ subtables” in this document.

pub fn seg_count_x2(&self) -> u16

2 × segCount.

pub fn search_range(&self) -> u16

Maximum power of 2 less than or equal to segCount, times 2 ((2floor(log2(segCount))) * 2, where “” is an exponentiation operator)

pub fn entry_selector(&self) -> u16

Log2 of the maximum power of 2 less than or equal to numTables (log2(searchRange/2), which is equal to floor(log2(segCount)))

pub fn range_shift(&self) -> u16

segCount times 2, minus searchRange ((segCount * 2) - searchRange)

pub fn end_code(&self) -> &'a [BigEndian<u16>]

End characterCode for each segment, last=0xFFFF.

pub fn start_code(&self) -> &'a [BigEndian<u16>]

Start character code for each segment.

pub fn id_delta(&self) -> &'a [BigEndian<i16>]

Delta for all character codes in segment.

pub fn id_range_offsets(&self) -> &'a [BigEndian<u16>]

Offsets into glyphIdArray or 0

pub fn glyph_id_array(&self) -> &'a [BigEndian<u16>]

Glyph index array (arbitrary length)

impl<'a> Cmap4<'a>

pub fn map_codepoint(&self, codepoint: impl Into<u32>) -> Option<GlyphId>

Maps a codepoint to a nominal glyph identifier.

pub fn iter(&self) -> Cmap4Iter<'a>

Returns an iterator over all (codepoint, glyph identifier) pairs in the subtable.

fn lookup_glyph_id( &self, codepoint: u16, index: usize, start_code: u16, ) -> Option<GlyphId>

Does the final phase of glyph id lookup.

Shared between Self::map and Cmap4Iter.

fn code_range(&self, index: usize) -> Option<Range<u32>>

Returns the [start_code, end_code] range at the given index.