read_fonts::tables::colr

Type Alias PaintSweepGradient

pub type PaintSweepGradient<'a> = TableRef<'a, PaintSweepGradientMarker>;

PaintSweepGradient table

Aliased Type

struct PaintSweepGradient<'a> { /* private fields */ }

Implementations

impl<'a> PaintSweepGradient<'a>

pub fn format(&self) -> u8

Set to 8.

pub fn color_line_offset(&self) -> Offset24

Offset to ColorLine table.

pub fn color_line(&self) -> Result<ColorLine<'a>, ReadError>

Attempt to resolve color_line_offset.

pub fn center_x(&self) -> FWord

Center x coordinate.

pub fn center_y(&self) -> FWord

Center y coordinate.

pub fn start_angle(&self) -> F2Dot14

Start of the angular range of the gradient, 180° in counter-clockwise degrees per 1.0 of value.

pub fn end_angle(&self) -> F2Dot14

End of the angular range of the gradient, 180° in counter-clockwise degrees per 1.0 of value.