Type Alias read_fonts::tables::gpos::MarkMarkPosFormat1

pub type MarkMarkPosFormat1<'a> = TableRef<'a, MarkMarkPosFormat1Marker>;

Mark-to-Mark Attachment Positioning Format 1: Mark-to-Mark Attachment

Aliased Type

struct MarkMarkPosFormat1<'a> { /* private fields */ }

Implementations

impl<'a> MarkMarkPosFormat1<'a>

pub fn pos_format(&self) -> u16

Format identifier: format = 1

pub fn mark1_coverage_offset(&self) -> Offset16

Offset to Combining Mark Coverage table, from beginning of MarkMarkPos subtable.

pub fn mark1_coverage(&self) -> Result<CoverageTable<'a>, ReadError>

Attempt to resolve mark1_coverage_offset.

pub fn mark2_coverage_offset(&self) -> Offset16

Offset to Base Mark Coverage table, from beginning of MarkMarkPos subtable.

pub fn mark2_coverage(&self) -> Result<CoverageTable<'a>, ReadError>

Attempt to resolve mark2_coverage_offset.

pub fn mark_class_count(&self) -> u16

Number of Combining Mark classes defined

pub fn mark1_array_offset(&self) -> Offset16

Offset to MarkArray table for mark1, from beginning of MarkMarkPos subtable.

pub fn mark1_array(&self) -> Result<MarkArray<'a>, ReadError>

Attempt to resolve mark1_array_offset.

pub fn mark2_array_offset(&self) -> Offset16

Offset to Mark2Array table for mark2, from beginning of MarkMarkPos subtable.

pub fn mark2_array(&self) -> Result<Mark2Array<'a>, ReadError>

Attempt to resolve mark2_array_offset.