Module read_fonts::tables::gvar

The gvar (Glyph Variations) table

Structs

DeltaIter

An iterator over the deltas for a glyph.

GlyphDelta

Delta information for a single point or component in a glyph.

GlyphVariationData

GvarFlags

TupleVariation

A single set of tuple variation data

TupleVariationIter

An iterator over the TupleVariations for a specific glyph.

U16Or32

Type Aliases

GlyphVariationDataHeader

The GlyphVariationData table

Gvar

The ‘gvar’ header

SharedTuples

Array of tuple records shared across all glyph variation data tables.