read_fonts::tables::gvar

Type Alias Gvar

pub type Gvar<'a> = TableRef<'a, GvarMarker>;

The ‘gvar’ header

Aliased Type

struct Gvar<'a> { /* private fields */ }

Implementations

impl<'a> Gvar<'a>

pub fn version(&self) -> MajorMinor

Major/minor version number of the glyph variations table — set to (1,0).

pub fn axis_count(&self) -> u16

The number of variation axes for this font. This must be the same number as axisCount in the ‘fvar’ table.

pub fn shared_tuple_count(&self) -> u16

The number of shared tuple records. Shared tuple records can be referenced within glyph variation data tables for multiple glyphs, as opposed to other tuple records stored directly within a glyph variation data table.

pub fn shared_tuples_offset(&self) -> Offset32

Offset from the start of this table to the shared tuple records.

pub fn shared_tuples(&self) -> Result<SharedTuples<'a>, ReadError>

Attempt to resolve shared_tuples_offset.

pub fn glyph_count(&self) -> u16

The number of glyphs in this font. This must match the number of glyphs stored elsewhere in the font.

pub fn flags(&self) -> GvarFlags

Bit-field that gives the format of the offset array that follows. If bit 0 is clear, the offsets are uint16; if bit 0 is set, the offsets are uint32.

pub fn glyph_variation_data_array_offset(&self) -> u32

Offset from the start of this table to the array of GlyphVariationData tables.

pub fn glyph_variation_data_offsets(&self) -> ComputedArray<'a, U16Or32>

Offsets from the start of the GlyphVariationData array to each GlyphVariationData table.

impl<'a> Gvar<'a>

pub fn data_for_gid(&self, gid: GlyphId) -> Result<FontData<'a>, ReadError>

pub fn glyph_variation_data( &self, gid: GlyphId, ) -> Result<GlyphVariationData<'a>, ReadError>

Get the variation data for a specific glyph.

pub fn phantom_point_deltas( &self, glyf: &Glyf<'_>, loca: &Loca<'_>, coords: &[F2Dot14], glyph_id: GlyphId, ) -> Result<[Point<Fixed>; 4], ReadError>

Returns the phantom point deltas for the given variation coordinates and glyph identifier.

The resulting array will contain four deltas: [left, right, top, bottom].