read_fonts::tables::hhea

Type Alias Hhea

pub type Hhea<'a> = TableRef<'a, HheaMarker>;

hhea Horizontal Header Table

Aliased Type

struct Hhea<'a> { /* private fields */ }

Implementations

impl<'a> Hhea<'a>

pub fn version(&self) -> MajorMinor

The major/minor version (1, 0)

pub fn ascender(&self) -> FWord

Typographic ascent.

pub fn descender(&self) -> FWord

Typographic descent.

pub fn line_gap(&self) -> FWord

Typographic line gap. Negative LineGap values are treated as zero in some legacy platform implementations.

pub fn advance_width_max(&self) -> UfWord

Maximum advance width value in ‘hmtx’ table.

pub fn min_left_side_bearing(&self) -> FWord

Minimum left sidebearing value in ‘hmtx’ table for glyphs with contours (empty glyphs should be ignored).

pub fn min_right_side_bearing(&self) -> FWord

Minimum right sidebearing value; calculated as min(aw - (lsb + xMax - xMin)) for glyphs with contours (empty glyphs should be ignored).

pub fn x_max_extent(&self) -> FWord

Max(lsb + (xMax-xMin))

pub fn caret_slope_rise(&self) -> i16

Used to calculate the slope of the cursor (rise/run); 1 for vertical caret, 0 for horizontal.

pub fn caret_slope_run(&self) -> i16

0 for vertical caret, 1 for horizontal.

pub fn caret_offset(&self) -> i16

The amount by which a slanted highlight on a glyph needs to be shifted to produce the best appearance. Set to 0 for non-slanted fonts

pub fn metric_data_format(&self) -> i16

0 for current format.

pub fn number_of_long_metrics(&self) -> u16

Number of LongMetric entries in ‘hmtx’/‘vmtx’ table