Module read_fonts::tables::layout

OpenType Layout common table formats

Structs

ClassRangeRecord

Used in ClassDefFormat2

FeatureRecord

Part of FeatureList

FeatureTableSubstitutionRecord

Used in FeatureTableSubstitution

FeatureVariationRecord

Part of FeatureVariations

LangSysRecord

LookupFlag

The LookupFlag bit enumeration.

RangeRecord

Used in CoverageFormat2

ScriptRecord

Script Record

SequenceLookupRecord

Sequence Lookup Record

Enums

ChainedSequenceContext

ClassDef

A Class Definition Table

Condition

Condition Table

CoverageTable

Coverage Table

DeltaFormat

Device delta formats

DeviceOrVariationIndex

Either a Device table (in a non-variable font) or a VariationIndex table (in a variable font)

FeatureParams

An enum for different possible tables referenced by Feature::feature_params_offset

SequenceContext

Subtables

an array of subtables, maybe behind extension lookups

Traits

ExtensionLookup

A trait that abstracts the behaviour of an extension subtable

Type Aliases

ChainedClassSequenceRule

Part of ChainedSequenceContextFormat2

ChainedClassSequenceRuleSet

Part of ChainedSequenceContextFormat2

ChainedSequenceContextFormat1

Chained Sequence Context Format 1

ChainedSequenceContextFormat2

Chained Sequence Context Format 2

ChainedSequenceContextFormat3

Chained Sequence Context Format 3

ChainedSequenceRule

Part of ChainedSequenceContextFormat1

ChainedSequenceRuleSet

Part of ChainedSequenceContextFormat1

CharacterVariantParams

featureParams for ‘cv01’-‘cv99’

ClassDefFormat1

Class Definition Table Format 1

ClassDefFormat2

Class Definition Table Format 2

ClassSequenceRule

Part of SequenceContextFormat2

ClassSequenceRuleSet

Part of SequenceContextFormat2

ConditionFormat1

Condition Table Format 1: Font Variation Axis Range

ConditionFormat2

Condition Table Format 2: Variation index

ConditionFormat3

Condition Table Format 3: AND

ConditionFormat4

Condition Table Format 4: OR

ConditionFormat5

Condition Table Format 5: NOT

ConditionSet

ConditionSet Table

CoverageFormat1

Coverage Format 1

CoverageFormat2

Coverage Format 2

Device

Device Table

Feature

Feature Table

FeatureList

Feature List Table

FeatureTableSubstitution

FeatureTableSubstitution Table

FeatureVariations

FeatureVariations Table

LangSys

Language System Table

Lookup

Lookup Table

LookupList

Lookup List Table

Script

Script Table

ScriptList

Script List Table

SequenceContextFormat1

Sequence Context Format 1

SequenceContextFormat2

Sequence Context Format 2

SequenceContextFormat3

Sequence Context Format 3

SequenceRule

Part of SequenceContextFormat1

SequenceRuleSet

Part of SequenceContextFormat1

SizeParams

StylisticSetParams

VariationIndex

Variation index table