read_fonts::tables::layout

Type Alias ChainedSequenceContextFormat3

pub type ChainedSequenceContextFormat3<'a> = TableRef<'a, ChainedSequenceContextFormat3Marker>;

Chained Sequence Context Format 3

Aliased Type

pub struct ChainedSequenceContextFormat3<'a> { /* private fields */ }

Implementations

impl<'a> ChainedSequenceContextFormat3<'a>

pub fn format(&self) -> u16

Format identifier: format = 3

pub fn backtrack_glyph_count(&self) -> u16

Number of glyphs in the backtrack sequence

pub fn backtrack_coverage_offsets(&self) -> &'a [BigEndian<Offset16>]

Array of offsets to coverage tables for the backtrack sequence

pub fn backtrack_coverages( &self, ) -> ArrayOfOffsets<'a, CoverageTable<'a>, Offset16>

A dynamically resolving wrapper for backtrack_coverage_offsets.

pub fn input_glyph_count(&self) -> u16

Number of glyphs in the input sequence

pub fn input_coverage_offsets(&self) -> &'a [BigEndian<Offset16>]

Array of offsets to coverage tables for the input sequence

pub fn input_coverages(&self) -> ArrayOfOffsets<'a, CoverageTable<'a>, Offset16>

A dynamically resolving wrapper for input_coverage_offsets.

pub fn lookahead_glyph_count(&self) -> u16

Number of glyphs in the lookahead sequence

pub fn lookahead_coverage_offsets(&self) -> &'a [BigEndian<Offset16>]

Array of offsets to coverage tables for the lookahead sequence

pub fn lookahead_coverages( &self, ) -> ArrayOfOffsets<'a, CoverageTable<'a>, Offset16>

A dynamically resolving wrapper for lookahead_coverage_offsets.

pub fn seq_lookup_count(&self) -> u16

Number of SequenceLookupRecords

pub fn seq_lookup_records(&self) -> &'a [SequenceLookupRecord]

Array of SequenceLookupRecords