read_fonts::tables::postscript

Module dict

Parsing for PostScript DICTs.

Structs

Blues

Operand for the BlueValues, OtherBlues, FamilyBlues and FamilyOtherBlues operators.

StemSnaps

Operand for the StemSnapH and StemSnapV operators.

Enums

Entry

PostScript DICT Operator with its associated operands.

Operator

PostScript DICT operator.

Token

Either a PostScript DICT operator or a (numeric) operand.

Functions

entries

Given a byte slice containing DICT data, returns an iterator yielding each operator with its associated operands.

tokens

Given a byte slice containing DICT data, returns an iterator yielding raw operands and operators.