read_fonts::tables::varc

Type Alias Varc

pub type Varc<'a> = TableRef<'a, VarcMarker>;

VARC (Variable Composites / Components Table)

FontTools VARC

Aliased Type

struct Varc<'a> { /* private fields */ }

Implementations

impl<'a> Varc<'a>

pub fn version(&self) -> MajorMinor

Major/minor version number. Set to 1.0.

pub fn coverage_offset(&self) -> Offset32

pub fn coverage(&self) -> Result<CoverageTable<'a>, ReadError>

Attempt to resolve coverage_offset.

pub fn multi_var_store_offset(&self) -> Nullable<Offset32>

pub fn multi_var_store( &self, ) -> Option<Result<MultiItemVariationStore<'a>, ReadError>>

Attempt to resolve multi_var_store_offset.

pub fn condition_list_offset(&self) -> Nullable<Offset32>

pub fn condition_list(&self) -> Option<Result<ConditionList<'a>, ReadError>>

Attempt to resolve condition_list_offset.

pub fn axis_indices_list_offset(&self) -> Nullable<Offset32>

pub fn axis_indices_list(&self) -> Option<Result<Index2<'a>, ReadError>>

Attempt to resolve axis_indices_list_offset.

pub fn var_composite_glyphs_offset(&self) -> Offset32

pub fn var_composite_glyphs(&self) -> Result<Index2<'a>, ReadError>

Attempt to resolve var_composite_glyphs_offset.

impl Varc<'_>

pub fn axis_indices(&self, nth: usize) -> Result<PackedDeltas<'_>, ReadError>

Friendlier accessor than directly using raw data via Index2

pub fn glyph(&self, nth: usize) -> Result<VarcGlyph<'_>, ReadError>

Friendlier accessor than directly using raw data via Index2

nth would typically be obtained by looking up a GlyphId in Self::coverage.