read_fonts::tables

Module variations

OpenType font variations common tables.

Structs

DeltaRunIter

Implements the logic for iterating over the individual runs

DeltaSetIndex

Outer and inner indices for reading from an ItemVariationStore.

EntryFormat

Entry format for a DeltaSetIndexMap.

FloatItemDelta

Floating point item delta computed by an item variation store.

PackedDeltas

Packed Deltas

PackedPointNumbers

Packed “Point” Numbers

PackedPointNumbersIter

An iterator over the packed point numbers data.

RegionAxisCoordinates

The RegionAxisCoordinates record

Tuple

A Tuple Record

TupleDeltaIter

An iterator over the deltas for a glyph.

TupleIndex

TupleVariation

A single set of tuple variation data

TupleVariationCount

The ‘tupleVariationCount’ field of the Tuple Variation Store Header

TupleVariationData

TupleVariationHeaderIter

A helper type for iterating over TupleVariationHeaders.

TupleVariationIter

An iterator over the TupleVariations for a specific glyph.

VariationRegion

The VariationRegion record

Enums

DeltaRunType

The type of values for a given delta run (influences the number of bytes per delta)

DeltaSetIndexMap

The DeltaSetIndexMap table

Constants

NO_VARIATION_INDEX

Traits

FloatItemDeltaTarget

Trait for applying floating point item deltas to target values.

TupleDelta

Trait for deltas that are computed in a tuple variation store.

Type Aliases

DeltaSetIndexMapFormat0

The DeltaSetIndexMap table format 0

DeltaSetIndexMapFormat1

The DeltaSetIndexMap table format 1

ItemVariationData

The ItemVariationData subtable

ItemVariationStore

The ItemVariationStore table

TupleVariationHeader

TupleVariationHeader

VariationRegionList

The VariationRegionList table