read_fonts

Trait TableProvider

pub trait TableProvider<'a> {

    // Required method
    fn data_for_tag(&self, tag: Tag) -> Option<FontData<'a>>;

    // Provided methods
    fn expect_data_for_tag(&self, tag: Tag) -> Result<FontData<'a>, ReadError> { ... }
    fn expect_table<T: TopLevelTable + FontRead<'a>>(
        &self,
    ) -> Result<T, ReadError> { ... }
    fn head(&self) -> Result<Head<'a>, ReadError> { ... }
    fn name(&self) -> Result<Name<'a>, ReadError> { ... }
    fn hhea(&self) -> Result<Hhea<'a>, ReadError> { ... }
    fn vhea(&self) -> Result<Vhea<'a>, ReadError> { ... }
    fn hmtx(&self) -> Result<Hmtx<'a>, ReadError> { ... }
    fn hdmx(&self) -> Result<Hdmx<'a>, ReadError> { ... }
    fn vmtx(&self) -> Result<Vmtx<'a>, ReadError> { ... }
    fn vorg(&self) -> Result<Vorg<'a>, ReadError> { ... }
    fn fvar(&self) -> Result<Fvar<'a>, ReadError> { ... }
    fn avar(&self) -> Result<Avar<'a>, ReadError> { ... }
    fn hvar(&self) -> Result<Hvar<'a>, ReadError> { ... }
    fn vvar(&self) -> Result<Vvar<'a>, ReadError> { ... }
    fn mvar(&self) -> Result<Mvar<'a>, ReadError> { ... }
    fn maxp(&self) -> Result<Maxp<'a>, ReadError> { ... }
    fn os2(&self) -> Result<Os2<'a>, ReadError> { ... }
    fn post(&self) -> Result<Post<'a>, ReadError> { ... }
    fn gasp(&self) -> Result<Gasp<'a>, ReadError> { ... }
    fn loca(
        &self,
        is_long: impl Into<Option<bool>>,
    ) -> Result<Loca<'a>, ReadError> { ... }
    fn glyf(&self) -> Result<Glyf<'a>, ReadError> { ... }
    fn gvar(&self) -> Result<Gvar<'a>, ReadError> { ... }
    fn cvt(&self) -> Result<&'a [BigEndian<i16>], ReadError> { ... }
    fn cvar(&self) -> Result<Cvar<'a>, ReadError> { ... }
    fn cff(&self) -> Result<Cff<'a>, ReadError> { ... }
    fn cff2(&self) -> Result<Cff2<'a>, ReadError> { ... }
    fn cmap(&self) -> Result<Cmap<'a>, ReadError> { ... }
    fn gdef(&self) -> Result<Gdef<'a>, ReadError> { ... }
    fn gpos(&self) -> Result<Gpos<'a>, ReadError> { ... }
    fn gsub(&self) -> Result<Gsub<'a>, ReadError> { ... }
    fn feat(&self) -> Result<Feat<'a>, ReadError> { ... }
    fn ltag(&self) -> Result<Ltag<'a>, ReadError> { ... }
    fn ankr(&self) -> Result<Ankr<'a>, ReadError> { ... }
    fn colr(&self) -> Result<Colr<'a>, ReadError> { ... }
    fn cpal(&self) -> Result<Cpal<'a>, ReadError> { ... }
    fn cblc(&self) -> Result<Cblc<'a>, ReadError> { ... }
    fn cbdt(&self) -> Result<Cbdt<'a>, ReadError> { ... }
    fn eblc(&self) -> Result<Eblc<'a>, ReadError> { ... }
    fn ebdt(&self) -> Result<Ebdt<'a>, ReadError> { ... }
    fn sbix(&self) -> Result<Sbix<'a>, ReadError> { ... }
    fn stat(&self) -> Result<Stat<'a>, ReadError> { ... }
    fn svg(&self) -> Result<Svg<'a>, ReadError> { ... }
    fn varc(&self) -> Result<Varc<'a>, ReadError> { ... }
    fn ift(&self) -> Result<Ift<'a>, ReadError> { ... }
    fn iftx(&self) -> Result<Ift<'a>, ReadError> { ... }
    fn meta(&self) -> Result<Meta<'a>, ReadError> { ... }

}

An interface for accessing tables from a font (or font-like object)

Required Methods

fn data_for_tag(&self, tag: Tag) -> Option<FontData<'a>>

Provided Methods

fn expect_data_for_tag(&self, tag: Tag) -> Result<FontData<'a>, ReadError>

fn expect_table<T: TopLevelTable + FontRead<'a>>(&self) -> Result<T, ReadError>

fn head(&self) -> Result<Head<'a>, ReadError>

fn name(&self) -> Result<Name<'a>, ReadError>

fn hhea(&self) -> Result<Hhea<'a>, ReadError>

fn vhea(&self) -> Result<Vhea<'a>, ReadError>

fn hmtx(&self) -> Result<Hmtx<'a>, ReadError>

fn hdmx(&self) -> Result<Hdmx<'a>, ReadError>

fn vmtx(&self) -> Result<Vmtx<'a>, ReadError>

fn vorg(&self) -> Result<Vorg<'a>, ReadError>

fn fvar(&self) -> Result<Fvar<'a>, ReadError>

fn avar(&self) -> Result<Avar<'a>, ReadError>

fn hvar(&self) -> Result<Hvar<'a>, ReadError>

fn vvar(&self) -> Result<Vvar<'a>, ReadError>

fn mvar(&self) -> Result<Mvar<'a>, ReadError>

fn maxp(&self) -> Result<Maxp<'a>, ReadError>

fn os2(&self) -> Result<Os2<'a>, ReadError>

fn post(&self) -> Result<Post<'a>, ReadError>

fn gasp(&self) -> Result<Gasp<'a>, ReadError>

fn loca(&self, is_long: impl Into<Option<bool>>) -> Result<Loca<'a>, ReadError>

is_long can be optionally provided, if known, otherwise we look it up in head.

fn glyf(&self) -> Result<Glyf<'a>, ReadError>

fn gvar(&self) -> Result<Gvar<'a>, ReadError>

fn cvt(&self) -> Result<&'a [BigEndian<i16>], ReadError>

Returns the array of entries for the control value table which is used for TrueType hinting.

fn cvar(&self) -> Result<Cvar<'a>, ReadError>

fn cff(&self) -> Result<Cff<'a>, ReadError>

fn cff2(&self) -> Result<Cff2<'a>, ReadError>

fn cmap(&self) -> Result<Cmap<'a>, ReadError>

fn gdef(&self) -> Result<Gdef<'a>, ReadError>

fn gpos(&self) -> Result<Gpos<'a>, ReadError>

fn gsub(&self) -> Result<Gsub<'a>, ReadError>

fn feat(&self) -> Result<Feat<'a>, ReadError>

fn ltag(&self) -> Result<Ltag<'a>, ReadError>

fn ankr(&self) -> Result<Ankr<'a>, ReadError>

fn colr(&self) -> Result<Colr<'a>, ReadError>

fn cpal(&self) -> Result<Cpal<'a>, ReadError>

fn cblc(&self) -> Result<Cblc<'a>, ReadError>

fn cbdt(&self) -> Result<Cbdt<'a>, ReadError>

fn eblc(&self) -> Result<Eblc<'a>, ReadError>

fn ebdt(&self) -> Result<Ebdt<'a>, ReadError>

fn sbix(&self) -> Result<Sbix<'a>, ReadError>

fn stat(&self) -> Result<Stat<'a>, ReadError>

fn svg(&self) -> Result<Svg<'a>, ReadError>

fn varc(&self) -> Result<Varc<'a>, ReadError>

fn ift(&self) -> Result<Ift<'a>, ReadError>

fn iftx(&self) -> Result<Ift<'a>, ReadError>

fn meta(&self) -> Result<Meta<'a>, ReadError>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.