Trait read_fonts::traversal::SomeRecord

pub trait SomeRecord<'a> {
    // Required method
    fn traverse(self, data: FontData<'a>) -> RecordResolver<'a>;
}

A generic trait for records, which need to be passed in data in order to fully resolve themselves.

Required Methods

fn traverse(self, data: FontData<'a>) -> RecordResolver<'a>