read_fonts::traversal

Trait SomeTable

pub trait SomeTable<'a> {
    // Required methods
    fn type_name(&self) -> &str;
    fn get_field(&self, idx: usize) -> Option<Field<'a>>;
}

A generic table type.

This is intended to be used as a trait object, and is a way of generically representing any table, providing ordered access to that table’s fields.

Required Methods

fn type_name(&self) -> &str

The name of this table

fn get_field(&self, idx: usize) -> Option<Field<'a>>

Access this table’s fields, in declaration order.

Implementations

impl<'a> dyn SomeTable<'a> + 'a

pub fn iter(&self) -> impl Iterator<Item = Field<'a>> + '_

Returns an iterator over this table’s fields.